% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{add_feature_counts}
\alias{add_feature_counts}
\title{Add Feature Counts}
\usage{
add_feature_counts(experiment, rnaseq_feature_counts,
  five_prime_feature_counts)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{rnaseq_feature_counts}{Raw counts in count matrix form from RNA-seq data}

\item{five_prime_feature_counts}{Raw counts in count matrix form from 5-prime sequencing data}
}
\description{
Helper function to add RNA-seq and 5prime feature count data.
}
