% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_groups.R
\name{add_feature_group}
\alias{add_feature_group}
\title{Feature Groups}
\usage{
add_feature_group(experiment, group_sheet, feature_type)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{group_sheet}{Group sheet as tab delimited file or data.frame
with column 'feature', and other named columns containing feature groups}

\item{feature_type}{either 'gene' or 'transcript'}
}
\description{
Add gene/transcript group info to tsrexplorer object
}
