% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{add_rna_seq}
\alias{add_rna_seq}
\title{Add RNAseq Data}
\usage{
add_rna_seq(experiment, rnaseq_count_matrix)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{rnaseq_count_matrix}{Raw counts in count matrix form from RNA-seq data}
}
\description{
Helper function to add RNA-seq data.
}
