% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{add_rnaseq_feature_counts}
\alias{add_rnaseq_feature_counts}
\title{Add RNAseq Data}
\usage{
add_rnaseq_feature_counts(experiment, rnaseq_feature_counts)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{rnaseq_count_matrix}{Raw counts in count matrix form from RNA-seq data}
}
\description{
Helper function to add RNA-seq data.
}
