% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{add_sample_sheet}
\alias{add_sample_sheet}
\title{Sample Sheet}
\usage{
add_sample_sheet(experiment, sample_sheet)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{sample_sheet}{Sample sheet as tab delimited file or data.frame.
Must have columns: 'sample_name' 'file_1' and 'file_2', and can have additional
  columns specifying condition, batch, etc.}
}
\description{
Add a sample sheet to a TSRexploreR object.
}
