% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{add_tss_feature_counts}
\alias{add_tss_feature_counts}
\title{Add TSS Total Counts}
\usage{
add_tss_feature_counts(experiment, tss_feature_counts)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{tss_total_count_matrix}{Raw counts in count matrix form from TSS mapping data}
}
\description{
Add TSS feature counts data.
}
