% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{annotate_differential}
\alias{annotate_differential}
\title{Annotate Differential TSSs or TSRs}
\usage{
annotate_differential(differential_exp, annotation_file,
  feature_type = c("gene", "transcript"), upstream = 1000,
  downstream = 100)
}
\arguments{
\item{differential_exp}{Tibble of differential TSSs or TSRs from differential_expression}

\item{annotation_file}{GTF genomic annotation file}

\item{feature_type}{Whether to annotate TSSs or TSRs relative to genes or transcripts}

\item{upstream}{Bases upstream of TSS}

\item{downstream}{Bases downstream of TSS}
}
\value{
Tibble of annotated differential TSSs or TSRs
}
\description{
Annotate Differential TSSs or TSRs to nearest gene or transcript.
}
