% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{annotate_differential_tsrs}
\alias{annotate_differential_tsrs}
\title{Annotate Differential TSRs}
\usage{
annotate_differential_tsrs(differential_tsrs, annotation_file,
  upstream = 1000, downstream = 100)
}
\arguments{
\item{differential_tsrs}{Tibble of differential TSRs from differential_tsrs}

\item{annotation_file}{GTF genomic annotation file}

\item{upstream}{Bases upstream of TSS}

\item{downstream}{Bases downstream of TSS}

\item{feature_type}{Whether to annotate TSRs relative to genes to transcripts}
}
\value{
Tibble of annotated differential TSRs
}
\description{
Annotate Differential TSRs to nearest gene or transcript.
}
