% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_features}
\alias{annotate_features}
\title{Annotate Data}
\usage{
annotate_features(experiment, annotation_file, data_type = c("tss",
  "tsr"), feature_type = c("gene", "transcript"), upstream = 1000,
  downstream = 100)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS Granges}

\item{annotation_file}{Either path and file name of annotation file, or TxDb object of annotation}

\item{data_type}{Whether to annotate TSSs or TSRs}

\item{feature_type}{Annotate on gene or transcript level}

\item{upstream}{Bases upstream of TSS}

\item{downstream}{Bases downstream of TSS}
}
\value{
Annotated TSS slot in tsrchitect object
}
\description{
Use ChIPseeker package to annotate TSSs or TSRs.
}
