% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_features}
\alias{annotate_features}
\title{Annotate Features}
\usage{
annotate_features(
  experiment,
  data_type = c("tss", "tsr", "tss_diff", "tsr_diff", "shift"),
  feature_type = c("gene", "transcript"),
  genome_annotation = NULL,
  upstream = 1000,
  downstream = 100
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to annotate TSSs ('tss') or TSRs ('tsr').}

\item{feature_type}{Whether to annotate at the 'gene' or 'transcript' level.}

\item{genome_annotation}{Genome annotation in GTF, GFF3, or TxDb format.}

\item{upstream}{Number of bases upstream of TSS for 'promoter' annotation (integer).}

\item{downstream}{Number of bases downstream of TSS for 'promoter' annotation (integer).}
}
\value{
TSRexploreR object with annotation data added to TSS or TSR tables.
}
\description{
Use the ChIPseeker package to annotate TSSs or TSRs relative to known genes or transcripts.
}
\details{
This function attempts to assign TSSs or TSRs to the nearest genomic feature.
Genomic annotation data can be provided as either a 'GTF' or 'GFF' file,
  or as a TxDb package from Bioconductor.

'feature_type' allows to you link TSSs or TSRs to genes or transcripts.
Furthermore, the size of the promoter region can be defined using
  'upstream' and 'downstream', which are relative to the TSSs
  defined in your annotation data.
TSSs or TSRs overlapping a gene on the opposite strand
  will be marked as 'Antisense'.
}
\examples{
data(TSSs)
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")

tsre <- TSSs[1] \%>\%
  tsr_explorer(genome_annotation=annotation) \%>\%
  format_counts(data_type="tss")

# Annotating TSSs.
tsre <- annotate_features(tsre, data_type="tss")

}
