% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply a threshold to TSSs or TSRs}
\usage{
apply_threshold(
  experiment,
  threshold,
  data_type = c("tss", "tsr"),
  use_normalized = FALSE
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}
}
\description{
Apply a threshold to TSSs or TSRs
}
