% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{associate_with_tsr}
\alias{associate_with_tsr}
\title{Associate TSSs}
\usage{
associate_with_tsr(experiment, use_sample_sheet = TRUE,
  sample_list = NA)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{use_sample_sheet}{Whether to associate TSSs with their matching TSR in the sample sheet}

\item{sample_list}{If 'use_sample_sheet' is FALSE, provide list with TSR as name and vector of TSSs to associate with it.}
}
\description{
Associate TSSs with TSRs
}
