% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{associate_with_tsr}
\alias{associate_with_tsr}
\title{Associate TSSs}
\usage{
associate_with_tsr(experiment, sample_list = NULL)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{sample_list}{List with TSRs as names and TSSs as vector.
If NULL will associate TSSs with TSRs from the sample of the same name.}
}
\description{
Associate TSSs with TSRs
}
\details{
TSRexploreR provides many options for the import and merging of TSSs and TSRs.
Because of this, TSS samples must be associated with TSR samples after TSR import,
  TSR merging, or TSS clustering using this function.
This adds an extra workflow step, but provides more analytical flexibility.
Each TSS with genomic coordinates overlapping those of a TSR in the specified TSR sample
  will be linked to that TSR.
TSSs not overlapping a TSR in the specified sample will not be associated with any TSR.

TSS samples can be associated with TSR samples using a list or sample sheet.
'sample_list' should be a named list of character vectors, with the names being the TSR
  sample names and the character vectors as the TSS samples(s) that should be associated with
  each TSR sample. If no sample list is provided, the function will automatically associate
  TSSs with the TSRs from the sample of the same name.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)

TSSs[1] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  associate_with_tsr

}
