% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{common_params}
\alias{common_params}
\title{Function to hold common parameters.}
\usage{
common_params(x)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{genome_assembly}{Genome assembly in either a FASTA file or BSgenome object format.}

\item{genome_annotation}{Genome annotation in either a GTF, GFF3, or TxDb object format.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}

\item{samples}{A vector of sample names to analyze.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{log2fc_cutoff}{Values below the |Log2 FC| cutoff will be discarded.}

\item{fdr_cutoff}{Values below the FDR cutoff will be discarded.}

\item{dominant}{If TRUE will discard any TSSs and/or TSRs that do not have the
highest score per gene or transcript.}

\item{data_conditions}{Apply advanced conditions to the data.}
}
\description{
Function to hold common parameters.
}
