% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{condition_data}
\alias{condition_data}
\title{Data Conditionals}
\usage{
condition_data(signal_data, data_conditions)
}
\arguments{
\item{signal_data}{TSS or TSR data.}

\item{data_conditions}{List of conditions.}
}
\description{
Apply strategies for filtering, ordering, quantiling, and/or grouping of data.
}
