% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{conditionals}
\alias{conditionals}
\title{Data Conditions Input}
\usage{
conditionals(
  data_filters = NULL,
  data_ordering = ordering(),
  data_quantiling = quantiling(),
  data_grouping = NULL
)
}
\arguments{
\item{data_filters}{Filter data by columns.}

\item{data_ordering}{Ordering object with order settings.
See ?ordering for more information.}

\item{data_quantiling}{Quantiling object with quantile settings.
See ?quantiling for more information.}

\item{data_grouping}{If quantiles not set,
split data by the specified categorical variable.}
}
\description{
Data Conditions Input
}
