% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{count_features}
\alias{count_features}
\title{Feature Counts}
\usage{
count_features(experiment, data_type = c("tss", "tsr"))
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Either 'tss' or 'tsr'}
}
\description{
Add feature counts
}
