% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{count_matrix}
\alias{count_matrix}
\title{Count Matrix}
\usage{
count_matrix(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"), samples = "all")
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{samples}{The samples to turn into a count matrix}
}
\value{
tsr explorer object with added count matrices
}
\description{
Generate count matrices for correlation analysis.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
tsre_exp <- count_matrix(tsre_exp, data_type = "tss")

}
\seealso{
\code{\link{tmm_normalize}} to normalize the count matrices.
  \code{\link{plot_correlation}} for various correlation plots.
}
