% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{cpm_normalize}
\alias{cpm_normalize}
\title{CPM Normalize Counts}
\usage{
cpm_normalize(experiment, data_type = c("tss", "tsr", "features"))
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{'tss', 'tsr', or 'features'}
}
\description{
CPM normalize the TSS, TSR, and/or feature counts.
}
