% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{de_table}
\alias{de_table}
\title{DE Table}
\usage{
de_table(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  de_comparisons = "all",
  de_type = c("up", "unchanged", "down")
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The name of the DE comparison}

\item{de_type}{A single value or combination of 'up, 'unchanged', and/or 'down' (qq a list?)}
}
\description{
Output a table with differential features
}
