% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{detect_features}
\alias{detect_features}
\title{Genes/Transcripts Detected}
\usage{
detect_features(experiment, samples = "all", data_type = c("tss",
  "tsr"), threshold = NA, dominant = FALSE, condition_data = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs or TSRs}

\item{samples}{Either 'all' or vector of sample names}

\item{data_type}{Whether TSSs or TSRs should be analyzed}

\item{threshold}{The number of reads required in a TSS or TSR to avoid filtering}

\item{condition_data}{Apply conditions to data (supports filtering, quantile, and/or grouping)}
}
\value{
tibble of detected feature numbers
}
\description{
Get the number of genes or transcripts detected
}
