% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Analyze Differential Features}
\usage{
differential_expression(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  comparison_name,
  comparison_type,
  comparison,
  shrink_lfc = FALSE
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether the input was generated from TSSs, TSRs, or genes/transcripts}

\item{comparison_name}{The name given to the comparison when stored back into the TSRexploreR object.}

\item{comparison_type}{For DEseq2, either 'contrast' or 'name'. For edgeR, either 'contrast' or 'coef'.}

\item{comparison}{For DESeq2, the contrast or name. For edgeR, the coefficients or contrasts.}

\item{shrink_lfc}{For DESeq2, whether the log2 fold changes are shrunk (TRUE) or not (FALSE).}
}
\value{
tibble of differential TSRs.
}
\description{
Find differential TSSs, TSRs, or features from edgeR or DESeq2 model.
}
