% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_tsrs}
\alias{differential_tsrs}
\title{Find Differential TSRs}
\usage{
differential_tsrs(fit_edger_model, comparisons = c())
}
\arguments{
\item{fit_edger_model}{edgeR differential expression model from fit_edger_model}

\item{comparison}{Vector of length two of the two groups to find differential TSRs}
}
\value{
tibble of differential TSRs
}
\description{
Find differential TSRs from edgeR model.
}
