% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_tsrs_volcano_plot}
\alias{differential_tsrs_volcano_plot}
\title{Differential TSR Volcano Plot}
\usage{
differential_tsrs_volcano_plot(differential_tsrs, log2fc_cutoff = 1,
  fdr_cutoff = 0.05)
}
\arguments{
\item{differential_tsrs}{Tibble of differential TSRs from differential_tsrs}

\item{log2fc_cutoff}{Log2 fold change cutoff for significance}

\item{fdr_cutoff}{FDR value cutoff for significance}
}
\value{
ggplot2 object of differential TSRs volcano plot.
}
\description{
Generate volcano plot for differential TSRs
}
