% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tss_plots.R
\name{dominant_tss}
\alias{dominant_tss}
\title{Annotate Dominant TSS}
\usage{
dominant_tss(experiment, threshold = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs}

\item{threshold}{Read threshold for TSS}
}
\value{
Tibble with dominant TSSs for each gene
}
\description{
Annotate Dominant TSS
}
