% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_clustering.R
\name{.aggr_scores}
\alias{.aggr_scores}
\title{Aggregate Scores}
\usage{
.aggr_scores(granges, maxdist, maxwidth)
}
\arguments{
\item{granges}{GRanges.}

\item{maxdist}{Maximum distance to cluster.}
}
\description{
Aggregate Scores
}
