% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotide.R
\name{.calculate_freqs}
\alias{.calculate_freqs}
\title{Internal DinucleotideFrequency Calculation}
\usage{
.calculate_freqs(seqs, grouping_status)
}
\arguments{
\item{seqs}{Dinucleotide sequences}

\item{grouping_status}{Whether any goruping should be applied}
}
\description{
Internal DinucleotideFrequency Calculation
}
