% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{.count_features}
\alias{.count_features}
\title{Calculate Feature Counts}
\usage{
.count_features(sample_data, grouping_status)
}
\arguments{
\item{sample_data}{Sample data.}

\item{grouping_status}{Whether there is a grouping variable.}
}
\description{
Calculate Feature Counts
}
