% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{.count_matrix}
\alias{.count_matrix}
\title{Generate Count Matrix.}
\usage{
.count_matrix(
  count_data,
  data_type = c("tss", "tsr"),
  use_normalized = FALSE,
  threshold = NULL
)
}
\arguments{
\item{count_data}{Named list of counts.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}
}
\description{
Generate Count Matrix.
}
