% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{.create_consensus}
\alias{.create_consensus}
\title{Create Consensus Ranges}
\usage{
.create_consensus(signal_data, conditions)
}
\arguments{
\item{signal_data}{TSS or TSR data}

\item{conditions}{Either ordering or quantiling conditions}
}
\description{
Create Consensus Ranges
}
