% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.de_status}
\alias{.de_status}
\title{Mark DE Status}
\usage{
.de_status(de_results, log2fc_cutoff, fdr_cutoff)
}
\arguments{
\item{de_results}{Results of DE}

\item{log2fc_cutoff}{Log2FC cutoff value}

\item{fdr_cutoff}{FDR cutoff value}
}
\description{
Mark DE Status
}
