% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.edger_model}
\alias{.edger_model}
\title{edgeR Differential Expression Model}
\usage{
.edger_model(count_data, sample_sheet, formula)
}
\arguments{
\item{count_data}{Count matrix}

\item{sample_sheet}{Sample data}

\item{formula}{Differential expression formula}
}
\description{
edgeR Differential Expression Model
}
