% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{.explore_thresholds}
\alias{.explore_thresholds}
\title{Threshold Exploration}
\usage{
.explore_thresholds(
  experiment,
  max_threshold = 25,
  steps = 0.5,
  samples = "all",
  use_normalized = FALSE
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{max_threshold}{Thresholds from 1 to max_threshold will be explored}

\item{steps}{Steps to get the threshold values}

\item{samples}{A vector of sample names to analyze.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}
}
\value{
data.frame containing information for each threshold and sample
}
\description{
Explore raw read count thresholds thresholds.
}
\details{
All TSS mapping technologies have some degree of spurious background reads.
tsr_explorer allows for basic thresholding of data by requiring
  a minimum number of reads for a TSS to be considered.

An easy way to pick a threshold is with the threshold plot generated
  by 'plot_exploration_threshold'.
This plot shows the proportion TSS positions that are promoter-proximal when
  various threshold values are tested.
'max_threshold' defines the maximum threshold value that will be analyzed.
}
\seealso{
\code{\link{plot_threshold_exploration}} to plot the results.
}
