% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{.filter_data}
\alias{.filter_data}
\title{Filtering}
\usage{
.filter_data(signal_data, filters)
}
\arguments{
\item{signal_data}{List of TSSs or TSRs}

\item{filters}{Quosure containing filters}
}
\description{
Filtering
}
