% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{.filter_heatmap}
\alias{.filter_heatmap}
\title{Filter Heatmap.}
\usage{
.filter_heatmap(sample_list, filtering)
}
\arguments{
\item{sample_list}{List of sample data.}

\item{filter}{Quosure of filters.}
}
\description{
Filter Heatmap.
}
