% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{.group_data}
\alias{.group_data}
\title{Group Data}
\usage{
.group_data(signal_data, conditions)
}
\arguments{
\item{signal_data}{TSS or TSR data}

\item{conditions}{List of conditions}
}
\description{
Group Data
}
