% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{.import_bigwigs}
\alias{.import_bigwigs}
\title{Import TSS bigwigs.}
\usage{
.import_bigwigs(sample_sheet)
}
\arguments{
\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}
}
\description{
Import TSS bigwigs.
}
