% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{.merge_overlapping_TSSs}
\alias{.merge_overlapping_TSSs}
\title{Merge Overlapping TSSs}
\usage{
.merge_overlapping_TSSs(samples, norm_status)
}
\arguments{
\item{samples}{Sample data}

\item{norm_status}{Whether there is a normalized score column.}
}
\description{
Merge Overlapping TSSs
}
