% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{.order_data}
\alias{.order_data}
\title{Ordering}
\usage{
.order_data(signal_data, ordering)
}
\arguments{
\item{signal_data}{List of TSSs or TSRs}

\item{ordering}{List of ordering parameters}
}
\description{
Ordering
}
