% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{.order_heatmap}
\alias{.order_heatmap}
\title{Order Heatmap.}
\usage{
.order_heatmap(
  count_data,
  data_type,
  annotated_data,
  ordering,
  order_fun,
  order_descending,
  order_samples
)
}
\arguments{
\item{count_data}{data.table of sample data.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{annotated_data}{Annotated sample data.}

\item{ordering}{Symbol/name specifying the column to order by.}

\item{order_fun}{Function to aggregate variable by before ordering.}

\item{order_descending}{Whether to order in descending (TRUE) order.}

\item{order_samples}{Samples that are used to calculate ordering.}
}
\description{
Order Heatmap.
}
