% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{.out_of_bounds_index}
\alias{.out_of_bounds_index}
\title{Get index of out of bounds ranges.}
\usage{
.out_of_bounds_index(x)
}
\arguments{
\item{x}{GenomicRanges object.}
}
\description{
Taken from GenomicRanges:::get_out_of_bound_index
}
