% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSRexplore.R
\name{.prepare_annotation}
\alias{.prepare_annotation}
\title{Prepare Genome Annotation}
\usage{
.prepare_annotation(annotation, experiment = NULL)
}
\arguments{
\item{annotation}{Genome annotation}

\item{experiment}{tsr explorer object}
}
\description{
Prepare Genome Annotation
}
