% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSRexplore.R
\name{.prepare_assembly}
\alias{.prepare_assembly}
\title{Prepare Genome Assembly}
\usage{
.prepare_assembly(assembly, experiment = NULL)
}
\arguments{
\item{assembly}{Genome assembly}

\item{experiment}{tsr explorer object}
}
\description{
Prepare Genome Assembly
}
