% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{.quantile_data}
\alias{.quantile_data}
\title{Quantiling}
\usage{
.quantile_data(signal_data, quantiling)
}
\arguments{
\item{signal_data}{List of TSSs or TSRs}

\item{quantiling}{List of quantiling parameters}
}
\description{
Quantiling
}
