% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{.quantile_heatmap}
\alias{.quantile_heatmap}
\title{Quantile Heatmap}
\usage{
.quantile_heatmap(
  count_data,
  data_type,
  annotated_data,
  quantiling,
  quantile_fun,
  n_quantiles,
  quantile_samples
)
}
\arguments{
\item{count_data}{data.table of sample data.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{annotated_data}{Annotated data.}

\item{quantiling}{Character specifying column to quantile by.}

\item{quantile_fun}{Functiont o aggregate variable by before quantiling.}

\item{n_quantiles}{Number of quantiles.}

\item{quantile_samples}{Samples to use for quantiling.}
}
\description{
Quantile Heatmap
}
