% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{.split_heatmap}
\alias{.split_heatmap}
\title{Split Heatmap}
\usage{
.split_heatmap(count_mat, split_by)
}
\arguments{
\item{count_mat}{Count matrix.}

\item{split_by}{Named list with split group as name and vector of genes,
or data.frame with columns 'feature' and 'split_group'.}
}
\description{
Split Heatmap
}
