% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{.tsr_heatmap}
\alias{.tsr_heatmap}
\title{TSR Heatmap Count Matrix}
\usage{
.tsr_heatmap(annotated_data, upstream = 1000, downstream = 1000)
}
\arguments{
\item{annotated_data}{Annotated data.table}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}
}
\value{
Matrix of counts for each gene/transcript and position
}
\description{
Generate count matrix to make TSR heatmap
}
