% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{.tss_association}
\alias{.tss_association}
\title{TSS Association}
\usage{
.tss_association(experiment, tss_names, tsr_name)
}
\arguments{
\item{experiment}{tsr explorer object}

\item{tss_names}{names of TSSs that will be ssociated with the TSR}

\item{tsr_name}{name of TSR that TSSs will be associated with}
}
\description{
TSS Association
}
