% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{explore_thresholds}
\alias{explore_thresholds}
\title{Threshold Exploration}
\usage{
explore_thresholds(experiment, max_threshold = 50, samples = "all")
}
\arguments{
\item{experiment}{tsrexplorer object with TSSs}

\item{max_threshold}{Threshold from 1 to max_threshold will be explored}

\item{samples}{Either 'all' or the names of the samples to analyze}
}
\value{
tibble containing information for each threshold and sample
}
\description{
Explore various naive thresholds
}
