% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{extract_de}
\alias{extract_de}
\title{Extract Differential Gene Sets}
\usage{
extract_de(experiment, data_type, de_comparisons = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The comparison sets to extract}
}
\description{
Extract differential expression analysis results.
}
