% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{find_correlation}
\alias{find_correlation}
\title{Generate Correlation Matrix}
\usage{
find_correlation(experiment, data_type = c("tss", "tsr", "features"),
  samples = "all", correlation_metric = "pearson")
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized counts}

\item{data_type}{Whether to make scatter plots from TSS, TSR, or RNA-seq & five-prime data}

\item{samples}{Either 'all' or the names of the samples to plot}

\item{correlation_metric}{Use either spearman or pearson correlation}
}
\value{
correlation matrix
}
\description{
Calculate a correlation matrix for sample concordance of TSSs or TSRs
}
