% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{find_correlation}
\alias{find_correlation}
\title{Generate Correlation Matrix}
\usage{
find_correlation(
  experiment,
  data_type = c("tss", "tsr", "features"),
  samples = "all",
  correlation_metric = "pearson"
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to make scatter plots from TSS, TSR, or RNA-seq & 5' data}

\item{samples}{A vector of sample names to analyze.}

\item{correlation_metric}{Whether to use Spearman or Pearson correlation}
}
\value{
Correlation matrix
}
\description{
Calculate a correlation matrix to assess between-sample concordance of TSSs or TSRs
}
