% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{fit_de_model}
\alias{fit_de_model}
\title{Model for differential feature analysis}
\usage{
fit_de_model(
  experiment,
  formula,
  data_type = c("tss", "tsr"),
  samples = "all",
  method = "DESeq2"
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{formula}{DE formula.}

\item{data_type}{Whether TSS, TSR, or gene/transcript counts should be analyzed.}

\item{samples}{A vector of sample names to analyze.}

\item{method}{Either 'DESeq2' or 'edgeR.}
}
\value{
TSRexploreR object with stored DE model.
}
\description{
Find differential TSSs or TSRs.
}
\details{
This function uses either DESeq2 or edgeR depending on what is specified in 'method'
  to find differential TSSs or TSRs.
'formula' should be a valid R formula in any form accepted by DESeq2 or edgeR,
  where the formula components are any columns present in the sample sheet.
}
\examples{
data(TSSs)
sample_sheet <- data.frame(
  sample_name=c(
    sprintf("S288C_D_\%s", seq_len(3)),
    sprintf("S288C_WT_\%s", seq_len(3))
  ),
  file_1=rep(NA, 6), file_2=rep(NA, 6),
  condition=c(
    rep("Diamide", 3),
    rep("Untreated", 3)
  )
)

tsre <- TSSs \%>\%
  tsr_explorer(sample_sheet=sample_sheet) \%>\%
  format_counts(data_type="tss")

# DESeq2 model for differential TSSs.
fit_de_model(tsre, ~condition, data_type="tss")

# DESeq2 model for differential TSRs.
tsre <- tss_clustering(tsre, threshold=3)
fit_de_model(tsre, ~condition, data_type="tsr")

}
\seealso{
\code{\link{differential_expression}} to extract differential TSSs or TSRs from model.
}
