% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{fit_edger_model}
\alias{fit_edger_model}
\title{Build edgeR Model of Differential TSRs}
\usage{
fit_edger_model(experiment, samples = c(), groups = c())
}
\arguments{
\item{experiment}{tsrexplorer objet after TMM normalization.}

\item{samples}{Vector of sample names to analyze.}

\item{groups}{Vector of groups in correct factor notation.}
}
\value{
DGEList object of differentially epxressed genes.with fitted DE model.
}
\description{
Find differential TSRs
}
