% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{format_counts}
\alias{format_counts}
\title{Format Counts}
\usage{
format_counts(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to format TSS or TSR counts}

\item{samples}{A vector of sample names to analyze.}
}
\value{
TSRexploreR object with properly formatted features
   in data.table format.
}
\description{
Format TSS or TSR counts for further analysis.
}
\details{
When TSSs or TSRs are first loaded into the TSRexploreR object
  they are stored as GRanges objects.
This function converts these into data.table format,
  and adds a few important columns for downstream analysis.
}
\examples{
library("magrittr")
data(TSSs)

TSSs[1] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss")

}
