% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{G_correction}
\alias{G_correction}
\title{Correct G}
\usage{
G_correction(experiment, assembly = NULL)
}
\arguments{
\item{experiment}{tsr explorer object}

\item{assembly}{Fasta assembly of genome or BSgenome object.}
}
\description{
Correct overrepresentation of G TSSs
}
