% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tracks.R
\name{gene_tracks}
\alias{gene_tracks}
\title{Gene Tracks by Gene}
\usage{
gene_tracks(experiment, genome_annotation, feature_name,
  feature_type = "gene", tss_samples = "all", tsr_samples = "all",
  threshold = 1, upstream = 250, downstream = 250,
  promoter_only = FALSE, use_cpm = FALSE, track_colors = "black",
  axis_scale = 0.25, ymax = NA)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{genome_annotation}{Genome annotation GTF/GFF file, or TxDb object}

\item{feature_name}{Name of gene or transcript to plot}

\item{feature_type}{Either 'gene' or 'transcript'}

\item{tss_samples}{Names of samples to plot TSSs for, or 'none'}

\item{tsr_samples}{Names of samples to plot TSRs for, or 'none'}

\item{threshold}{TSSs and TSRs below threshold are excluded from plotting}

\item{upstream}{bases upstream to extend gene or promoter track}

\item{downstream}{bases downstream to extend gene or promoter track}

\item{promoter_only}{Instead of plotting the entire gene, plot the promoter region}

\item{use_cpm}{Use CPM normalized reads or not}

\item{track_colors}{Either a single color value for all tracks, or a vector of colors}

\item{axis_scale}{Relative size scale for axis text and title}

\item{ymax}{Maximum value on Y axis for all TSS tracks}
}
\description{
Generate gene tracks in GViz by gene name.
}
