% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Genomic Distribution}
\usage{
genomic_distribution(experiment, data_type = c("tss", "tsr"),
  samples = "all", threshold = NA, dominant = FALSE,
  data_group = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSRs}

\item{data_type}{Whether to get distribution of TSSs or TSRs}

\item{samples}{Either "all" or vector of sample names}

\item{threshold}{Filter out TSSs or TSRs under a certain read count number}

\item{dominant}{Should dominant TSS per 'gene' or 'tsr' only be considered}

\item{data_group}{List of group options (filtering and quantiles available)}
}
\value{
tibble with TSS or TSR genomic distribution stats
}
\description{
Get genomic distribution of TSSs and TSRs
}
