% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_annotation}
\alias{get_annotation}
\title{Get Genome Annotation}
\usage{
get_annotation(experiment)
}
\arguments{
\item{experiment}{TSRexploreR object.}
}
\value{
TxDb object of genome annotation.
}
\description{
Extract the genome annotation.
}
\examples{
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")
tsre <- tsr_explorer(genome_annotation=annotation)
get_annotation(tsre)

}
\seealso{
\code{\link{get_assembly}} to get genome assembly.
}
