% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_counts}
\alias{get_counts}
\title{Get TSS or TSR count tables.}
\usage{
get_counts(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Get either TSS ('tss') or TSR ('tsr') count tables.}

\item{samples}{A vector of sample names to analyze.}
}
\value{
List of TSS or TSR count data.frame.
}
\description{
Extract the TSS or TSR count tables from the 'counts' slot.
}
\examples{
data(TSSs_reduced)

exp <- tsr_explorer(TSSs_reduced) \%>\%
  format_counts(data_type="tss")

cts <- get_counts(exp)

}
\seealso{
\code{\link{get_granges}} to get TSS or TSR GRanges.
}
