% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_diff_model}
\alias{get_diff_model}
\title{Get Differential TSS/TSR models}
\usage{
get_diff_model(experiment, data_type = c("tss", "tsr"))
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}
}
\value{
DESeq2 or edgeR differential expression model.
}
\description{
Get the DESeq2 or edgeR model.
}
