% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_diff_results}
\alias{get_diff_results}
\title{Get Differential TSS/TSR results}
\usage{
get_diff_results(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{samples}{A vector of sample names to analyze.}
}
\value{
List of differential features data.frames.
}
\description{
Get the DESeq2 or edgeR differential features.
}
