% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_sample_sheet}
\alias{get_sample_sheet}
\title{Get Sample Sheet}
\usage{
get_sample_sheet(experiment)
}
\arguments{
\item{experiment}{TSRexploreR object.}
}
\value{
data.frame
}
\description{
Extract the sample sheet.
}
\examples{
sample_sheet <- data.frame(
  sample_name=sprintf("S288C_WT_\%s", seq_len(3)),
  file_1=NA, file_2=NA,
  condition=rep("Untreated", 3)
)

exp <- tsr_explorer(sample_sheet=sample_sheet)

ss <- get_sample_sheet(exp)

}
