% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_data}
\alias{group_data}
\title{Group Data}
\usage{
group_data(signal_data, filter_by = NA, upper_filter = NA,
  lower_filter = NA, order_by = NA, order_direction = "descending",
  quantile_by = NA, n_quantiles = NA)
}
\arguments{
\item{signal_data}{TSS or TSR data}

\item{filter_by}{Continuous variable to filter the data by}

\item{upper_filter}{Upper value to filter continuous data by}

\item{lower_filter}{Lower value to filter continuous data by}

\item{order_by}{Metric to order data by}

\item{order_direction}{Whether the values should be ordered in
'ascending' or 'descending' order}

\item{quantile_by}{Continuous metric to calculate quantiles}

\item{n_quantiles}{Number of quantiles to calculate for continuous metric}
}
\description{
Group data for plotting
}
