% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tss_plots.R
\name{mark_dominant}
\alias{mark_dominant}
\title{Mark Dominant}
\usage{
mark_dominant(experiment, data_type = c("tss", "tsr"), threshold = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs/TSRs}

\item{data_type}{Either 'tss' or 'tsr'}

\item{threshold}{Read threshold for TSS/TSRs}
}
\value{
Tibble with dominant TSSs for each gene
}
\description{
Mark Dominant
}
