% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_metrics.R
\name{mark_dominant}
\alias{mark_dominant}
\title{Mark Dominant}
\usage{
mark_dominant(
  experiment,
  data_type = c("tss", "tsr"),
  threshold = NULL,
  use_normalized = FALSE,
  mark_per = "default",
  exclude_antisense = TRUE
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to mark dominant TSSs or TSRs.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{mark_per}{By default, the function marks dominant TSR per gene, and dominant TSS per TSR.
TSSs can also be set per as dominant TSS per 'gene'.}

\item{exclude_antisense}{Remove antisense TSSs/TSRs prior to analysis.}
}
\value{
TSRexploreR object with dominant status added to TSSs or TSRs.
}
\description{
Mark TSSs as dominant per TSR or gene/transcript, or TSRs as dominant per gene/transcript.
}
\details{
This function marks which TSSs are dominant per TSR or gene, or which TSR is
dominant per gene. Analysis of dominant features may help to cut through the 
noise to get information such as the length of the primary 5' UTR and sequence
features associated with the the strongest TSS.

Setting a 'threshold' will only mark a TSS or TSR as dominant if its score 
is greater than or equal to the threshold.

'mark_per' controls the behavior of the function. For TSSs 'default' will mark,
dominant TSS per TSR, and for TSRs the dominant TSR per gene is marked. For TSSs, 
'gene' can also be specified, which will mark the dominant TSS per gene.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")

tsre <- TSSs[1] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%

# Dominant TSS per gene.
tsre <- annotate_features(tsre, data_type="tss")
\donttest{mark_dominant(tsre, data_type="tss", mark_per="gene")}

# Dominant TSS per TSR.
tsre <- tsre \%>\%
  tss_clustering(threshold=3) \%>\%
  associate_with_tsr
\donttest{mark_dominant(tsre, data_type="tss")}

# Dominant TSR per gene.
tsre <- annotate_features(tsre, data_type="tsr")
\donttest{mark_dominant(tsre, data_type="tsr")}

}
\seealso{
\code{\link{associate_with_tsr}} to associate TSSs with TSRs prior to marking
  dominant TSS per TSR.
}
