% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge Samples}
\usage{
merge_samples(
  experiment,
  data_type = c("tss", "tsr"),
  threshold = NULL,
  sample_sheet = NULL,
  merge_group = NULL,
  merge_list = NULL,
  max_distance = NULL,
  genome_assembly = NULL
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{merge_group}{Column in sample sheet to merge by.}

\item{merge_list}{Named list of samples to merge.}

\item{max_distance}{Merge TSRs within this distance.}

\item{genome_assembly}{Genome assembly in FASTA or BSgenome format.}

\item{merge_replicates}{If 'TRUE', replicate groups will be merged.}

\item{sample_list}{If merge_replicates is set to 'FALSE',
specify what samples to merge in list format.}
}
\description{
Merge replicates or selected samples.
}
