% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,tsr_object-method}
\title{TMM Normalization}
\usage{
normalize(experiment)

\S4method{normalize}{tsr_object}(experiment)
}
\arguments{
\item{experiment}{tsrexplorer object}
}
\value{
tibble of TMM normalized read counts
}
\description{
Using edgeR to TMM normalize TSSs
}
