% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_counts}
\alias{normalize_counts}
\title{TMM Normalize TSSs or TSRs}
\usage{
normalize_counts(
  experiment,
  data_type = c("tss", "tss_features"),
  method = "DESeq2",
  threshold = 1,
  n_samples = 1
)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Whether TSSs, TSRs, or RNA-seq & 5' feature counts should be normalized}

\item{threshold}{Consider only features with at least this number of raw counts}

\item{n_samples}{Filter out positions without features meeting the the selected threshold
in this number of samples}

\item{normalization_method}{Either 'edgeR', 'DESeq2', or 'CPM'}
}
\value{
tsrexplorer object with tmm normalized count matrices
}
\description{
Using edgeR to TMM normalize TSSs or TSRs.
}
\details{
The TMM normalization method, employed by edgeR, is designed to reduce the influence of
  library size on inter-sample comparisons.

For TMM normalization, it is recommended to remove features with few or no reads as
  these may bias the final results.
To facilitate this filtering, two arguments are provided, 'threshold' and 'n_samples'.
Features must have greater than or equal to 'threshold' number of raw counts in at least
  'n_samples' number of samples to be retained.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- count_matrix(tsre_exp, data_type="tss")
tsre_exp <- tmm_normalize(exp, data_type="tss")

}
\seealso{
\code{\link{count_matrix}} to prepare the matrices.
  \code{\link{plot_correlation}} for various correlation plots.
}
