% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize TSS counts}
\usage{
normalize_counts(
  experiment,
  data_type = c("tss", "tss_features"),
  method = "DESeq2",
  threshold = 1,
  n_samples = 1
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether TSS or gene/transcript counts should be normalized.}

\item{method}{Either 'edgeR', 'DESeq2', or 'CPM'.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{n_samples}{Filter out TSSs or features not meeting the the selected threshold
in this number of samples.}
}
\value{
TSRexploreR object with normalized counts.
}
\description{
edgeR, DESeq2, or CPM normalization of TSSs or gene/transcript counts.
}
\details{
This function performs one of three normalizations on TSS or gene/transcript counts. The
simplest of these is counts per million (CPM), which accounts for sequencing depth. While
CPM is appropriate for comparing replicates, it is considered to be too simple for cases
in which there are expected to be substantial differences in RNA composition between samples.
For between-sample comparisons, the trimmed median of M-values (TMM) or median-of-ratios (MOR)
approaches, implemented in edgeR and DESeq2, respectively, can be used. Both of these methods
are designed to reduce the impact of library size on such comparisons. Prior to TMM or MOR
normalization, it is recommend to remove features with few or no reads, as they may bias the 
final results. To facilitate this filtering, two arguments are provided: 'threshold' and 
'n_samples'. Features must have greater than or equal to 'threshold' number of raw counts 
in at least 'n_samples' number of samples to proceed through normalization.

When clustering TSSs into TSRs using 'tss_clustering',
  both the raw and normalized counts will be stored in the new TSRs.
}
\examples{
library("magrittr")
data(TSSs)
sample_sheet <- data.frame(
  sample_name=c(
    sprintf("S288C_D_\%s", seq_len(3)),
    sprintf("S288C_WT_\%s", seq_len(3))
  ),
  file_1=NA, file_2=NA,
  condition=c(
    rep("Diamide", 3),
    rep("Untreated", 3)
  )
)

tsre <- TSSs \%>\%
  tsr_explorer(sample_sheet=sample_sheet) \%>\%
  format_counts(data_type="tss")

# CPM normalization
\donttest{normalize_counts(tsre, method="CPM")}

# DESeq2 normalization
\donttest{normalize_counts(tsre, method="DESeq2")}

# edgeR normalization
\donttest{normalize_counts(tsre, method="edgeR")}

}
