% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{ordering}
\alias{ordering}
\title{Ordering}
\usage{
ordering(..., .samples = NULL, .aggr_fun = mean)
}
\arguments{
\item{...}{Variables to order by.
Wrap varaible name in desc() for descending order (like in dplyr::arrange).}

\item{samples}{Names of samples to order by aggregate score.}

\item{If}{more than one sample is selected feature values
are aggregated using this function.}
}
\description{
Ordering
}
