% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_metrics.R
\name{peak_concentration}
\alias{peak_concentration}
\title{Peak Concentration}
\usage{
peak_concentration(tss_table)
}
\arguments{
\item{tss_table}{data.table of TSSs}
}
\description{
Calculate peak concentration.
}
