% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plots.R
\name{plot_average}
\alias{plot_average}
\title{Average Plots}
\usage{
plot_average(experiment, data_type = c("tss", "tsr"), samples = "all",
  consider_score = FALSE, upstream = 1000, downstream = 1000,
  threshold = 1, ncol = 1, use_cpm = FALSE, dominant = FALSE,
  data_conditions = NA, ...)
}
\arguments{
\item{experiment}{tsr_explorer object with annotated TSSs}

\item{data_type}{Make average plot for TSS or TSR data}

\item{samples}{Either 'all' to plot all samples, or a vector of sample names}

\item{consider_score}{Should the TSS or TSR score be considered, or just the unique location.}

\item{upstream}{Bases upstream of plot center}

\item{downstream}{Bases downstream of plot center}

\item{threshold}{threshold value for TSSs}

\item{ncol}{Number of columns to use when plotting data when quantiles not set}

\item{use_cpm}{Whether to use the CPM normalized or raw counts}

\item{dominant}{Consider only dominant TSS or TSR}

\item{data_conditions}{Data conditioning filters}

\item{...}{Arguments passed to geom_density}
}
\value{
ggplot2 object of average plot
}
\description{
Generate average plots of TSSs or TSRs
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
plot_average(tsre_exp, data_type = "tss")

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
}
