% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotide.R
\name{plot_dinucleotide_freqs}
\alias{plot_dinucleotide_freqs}
\title{Plot Dinucleotide Frequencies}
\usage{
plot_dinucleotide_freqs(dinucleotide_frequencies, ncol = 1, ...)
}
\arguments{
\item{dinucleotide_frequencies}{tibble from dinucleotide_frequencies analysis}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{...}{Arguments passed to geom_col}
}
\value{
ggplot2 object of dinucleotide frequencies plot
}
\description{
Plot results from dinucleotide analysis
}
\details{
This plotting function returns a ggplot2 barplot of -1 and +1 dinucleotide frequencies, 
  where +1 is the TSS and -1 is the position immediately upstream.
The results of the 'dinucleotide_frequencies' function are used as input.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
assembly <- system.file("extdata", "S288C_Assembly.fasta", package="TSRexploreR")
freqs <- dinucleotide_frequencies(tsre_exp, genome_assembly=assembly)
plot_dinucleotide_frequencies(freqs)

}
\seealso{
\code{\link{dinucleotide_frequencies}} to calculate the dinucleotide frequencies.
}
