% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotide.R
\name{plot_dinucleotide_frequencies}
\alias{plot_dinucleotide_frequencies}
\title{Plot Dinucleotide Frequencies}
\usage{
plot_dinucleotide_frequencies(dinucleotide_frequencies, ncol = 1, ...)
}
\arguments{
\item{dinucleotide_frequencies}{tibble from dinucleotide_frequencies analysis}

\item{ncol}{Number of columns to plot if not quantile plot}

\item{...}{Arguments passed to geom_col}
}
\value{
ggplot2 object of dinucleotide frequencies plot
}
\description{
Plot results from dinucleotide analysis
}
\details{
This plotting function returns a ggplot2 barplot of the +1 and -1 base
  frequencies (corresponding to the TSS and adjacent upstream base respectively)
  surrounding TSSs.
The results of the 'dinucleotide_frequencies' function are used as input.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
assembly <- system.file("extdata", "S288C_Assembly.fasta", package = "tsrexplorer")
freqs <- dinucleotide_frequencies(tsre_exp, genome_assembly = assembly)
plot_dinucleotide_frequencies(freqs)

}
\seealso{
\code{\link{dinucleotide_frequencies}} to calculate the dinucleotide frequencies.
}
