% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tss_plots.R
\name{plot_dominant_tss}
\alias{plot_dominant_tss}
\title{Plot Dominant TSS}
\usage{
plot_dominant_tss(experiment, upstream = 2000, downstream = 500,
  ncol = 1, consider_score = FALSE, quantiles = NA, ...)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{upstream}{Bases upstream to display on plot}

\item{downstream}{Bases downstream to display on plot}

\item{ncol}{Number of columns to plot the data to}

\item{consider_score}{Consider not just unique positions but the score of the TSS also}

\item{quantiles}{Break down data into quantiles}

\item{...}{Arguments passed to geom_density}

\item{samples}{Samples to plot}
}
\value{
ggplot2 object dominant TSS plot
}
\description{
Plot Dominant TSS
}
