% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(heatmap_matrix, max_value = 5, ncol = 1, ...)
}
\arguments{
\item{heatmap_matrix}{TSS or TSR heatmap matrix from tss_heatmap_matrix ot tsr_heatmap_matrix}

\item{max_value}{Max log2 value to truncate heatmap color}

\item{ncol}{Number of columns when plotting multiple samples}

\item{...}{Arguments passed to geom_tile}
}
\value{
ggplot2 object of TSS heatmap
}
\description{
Plot heatmap from count matrix generated by tss_heatmap_matrix or tsr_heatmap_matrix
}
