% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tss_plots.R
\name{plot_max_utr}
\alias{plot_max_utr}
\title{Plot Max UTR Length}
\usage{
plot_max_utr(max_utr, upstream = 1000, downstream = 100)
}
\arguments{
\item{max_utr}{tibble of max UTRs output by max_utr}

\item{upstream}{Bases upstream to extend average to}

\item{downstream}{Bases downstream to extend average to}
}
\value{
ggplot2 object of max UTR length average
}
\description{
Plot TSS with furthest distance
}
