% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_num_de}
\alias{plot_num_de}
\title{Plot DE Numbers}
\usage{
plot_num_de(
  experiment,
  data_type = c("tss", "tsr"),
  de_comparisons = "all",
  log2fc_cutoff = 1,
  fdr_cutoff = 0.05,
  keep_unchanged = FALSE,
  return_table = FALSE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{de_comparisons}{Character vector of differential expression comparisons to plot.}

\item{log2fc_cutoff}{Differential features not meeting this |Log2 FC| threshold will not be considered.}

\item{fdr_cutoff}{Differential features not meeting this significance threshold will not be considered.}

\item{keep_unchanged}{Whether to include (TRUE) unchanged features in the plot.}

\item{return_table}{Return a table of results instead of a plot.}

\item{...}{Additional arguments passed to geom_col.}
}
\value{
ggplot2 object of stacked barplot.
  If 'return_table' is TRUE, a data.frame with differentially expressed TSS/TSR numbers
  are returned.
}
\description{
Stacked barplot of the number of differential TSSs or TSRs per comparison.
}
\details{
Generate a stacked barplot with the number of differential TSSs or TSRs per comparison.

'de_comparisons' are the names given to the comparisons from the 'comparison_name'
  argument of the 'differential_expression' function.
'log2fc_cutoff' and 'fdr_cutoff' are the Log2 Fold Change and FDR cutoffs used
  for consideration of significance in the plot.
'keep_unchaged' controls whether non-significant feature numbers are included in the plot.

If 'keep_unchaged' is TRUE, a table with the numbers is returned instead of the ggplot.
This may be useful if exact numbers underlying the plot are required.
}
\examples{
data(TSSs)
sample_sheet <- data.frame(
  sample_name=c(
    sprintf("S288C_D_\%s", seq_len(3)),
    sprintf("S288C_WT_\%s", seq_len(3))
  ),
  file_1=NA, file_2=NA,
  condition=c(rep("Diamide", 3), rep("Untreated", 3))
)

tsre <- TSSs \%>\%
  tsr_explorer(sample_sheet=sample_sheet) \%>\%
  format_counts(data_type="tss")

# Differential TSS quantities plot.
diff_tss <- tsre \%>\%
  fit_de_model(data_type="tss", formula= ~condition) \%>\%
  differential_expression(
    data_type="tss",
    comparison_name="Diamide_vs_Untreated",
    comparison_type="name",
    comparison="condition_Untreated_vs_Diamide"
  )
\donttest{plot_num_de(diff_tss, data_type="tss")}

# Differential TSR quantities plot.
diff_tsr <- tsre \%>\%
  tss_clustering(threshold=3) \%>\%
  fit_de_model(data_type="tsr", formula= ~condition) \%>\%
  differential_expression(
    data_type="tsr",
    comparison_name="Diamide_vs_Untreated",
    comparison_type="name",
    comparison="condition_Untreated_vs_Diamide"
  )
\donttest{plot_num_de(diff_tsr, data_type="tsr")}

}
\seealso{
\code{\link{fit_de_model}} to fit a differential expression model.
\code{\link{differential_expression}} to find differential TSSs or TSRs.
}
