% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Dimensionality Reduction.}
\usage{
plot_reduction(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"), method = "umap", n_neighbors = 2, min_dist = 0.1,
  ...)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{method}{'umap' or 'pca'}

\item{n_neighbors}{Nearest neighbors parameter for umap}

\item{min_dist}{Minimum distance between points parameter for umap}

\item{...}{Additional arguments passed to geom_point}
}
\description{
Dimensionality Reduction.
}
