% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_rnaseq_corr}
\alias{plot_rnaseq_corr}
\title{RNA-seq TSSs total Correlation Matrix}
\usage{
plot_rnaseq_corr(experiment, corr_metrix = c("pearson", "spearman"))
}
\arguments{
\item{experiment}{tsrexplorer object with RNA-seq and TSS total added}

\item{corr_metric}{Correlation metrix to use ("pearon", "spearman")}
}
\value{
ggplot2 object of RNA-seq versus TSS mapping.
}
\description{
Correlation between total TSSs and RNA-seq
}
