% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{plot_sequence_logo}
\alias{plot_sequence_logo}
\title{Generate Sequence Logo}
\usage{
plot_sequence_logo(tss_sequences)
}
\arguments{
\item{tss_sequences}{Sequences surrounding TSS generated with tss_sequences}
}
\value{
ggplot2 object with sequence logo
}
\description{
Create a sequence logo for the sequences around TSSs
}
