% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_rank}
\alias{plot_shift_rank}
\title{Shifting Rank Plot}
\usage{
plot_shift_rank(
  experiment,
  samples = "all",
  score_order = "descending",
  ncol = 3
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{score_order}{Either descending or ascending}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}
}
\description{
Shifting Rank Plot
}
