% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{plot_threshold_exploration}
\alias{plot_threshold_exploration}
\title{Plot Threshold Exploration}
\usage{
plot_threshold_exploration(
  experiment,
  max_threshold = 25,
  steps = 1,
  samples = "all",
  use_normalized = FALSE,
  ncol = 1,
  point_size = 1,
  return_table = FALSE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{max_threshold}{Thresholds from 1 to max_threshold will be explored}

\item{steps}{Steps to get the threshold values}

\item{samples}{A vector of sample names to analyze.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{point_size}{The size of the points on the plot}

\item{return_table}{Return a table of results instead of a plot.}

\item{...}{Arguments passed to geom_point}
}
\value{
ggplot2 object containing the threshold exploration plot
}
\description{
Make a plot to explore threshold values.
}
\details{
We have found that a threshold of 3 often provides an appropriate balance between
  a high promoter-proximal fraction (>= 0.8) and the number of unique genes or
  transcripts with at least one unique TSS.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data=annotation,
  data_type="tss", feature_type="transcript"
)
thresh_results <- explore_thresholds(tsre_exp)
plot_threshold_exploration(thresh_results)

}
\seealso{
\code{\link{explore_thresholds}} for initial calculations.
}
