% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plots.R
\name{plot_tsr_average}
\alias{plot_tsr_average}
\title{TSR Average Plots}
\usage{
plot_tsr_average(experiment, sample, upstream = 1000,
  downstream = 1000)
}
\arguments{
\item{experiment}{tsr_explorer object with annotated TSRs}

\item{upstream}{Bases upstream of plot center}

\item{downstream}{Bases downstream of plot center}
}
\value{
ggplot2 object of average plot
}
\description{
Generate average plot of TSRs versus annotated TSSs
}
