% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_tsr_corr}
\alias{plot_tsr_corr}
\title{Replicate TSR Correlation Matrix}
\usage{
plot_tsr_corr(experiment, corr_metric = c("pearson", "spearman"))
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized TSR counts}

\item{corr_metric}{Correlation metric ("pearson", "spearman")}
}
\value{
ggplot2 object
}
\description{
Correlation matrix to explore replicate concordance.
}
