% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{plot_tsr_genomic_distribution}
\alias{plot_tsr_genomic_distribution}
\title{Plot TSR Genomic Distribution}
\usage{
plot_tsr_genomic_distribution(tsr_genomic_distribution)
}
\arguments{
\item{tsr_genomic_distribution}{tibble of TSR genomic distributions from tsr_genomic_distribution}
}
\value{
ggplot2 object with TSR genomic distribution plot
}
\description{
Get genomic distribution of TSRs
}
