% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_tsr_heatmap}
\alias{plot_tsr_heatmap}
\title{Plot TSR Heatmap}
\usage{
plot_tsr_heatmap(tsr_count_matrix, max_value = 8)
}
\arguments{
\item{tsr_count_matrix}{TSS count matrix from tss_count_matrix}

\item{max_value}{Max log2 value to truncate heatmap color}
}
\value{
ggplot2 object of TSS heatmap
}
\description{
Plot TSR heatmap from count matrix generated by tsr_count_matrix
}
