% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_plots.R
\name{plot_tsr_metric}
\alias{plot_tsr_metric}
\title{Plot TSR Stats}
\usage{
plot_tsr_metric(experiment, sample, metric, log2_transform = FALSE,
  xlims = c())
}
\arguments{
\item{experiment}{tsrexplorer object with TSR granges}

\item{sample}{Name of sample to return plot for}

\item{metric}{Name of metric in tsrexplorer TSR granges to plot}

\item{log2_transform}{Whether the metric should be log2+1 transformed prior to plotting}

\item{xlims}{Vector of values specifying the upper and lower limits of the x-axis}
}
\value{
ggplot2 object with tsr metrix plotted
}
\description{
Plot selected TSR stats
}
