% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_tsr_scatter}
\alias{plot_tsr_scatter}
\title{Replicate TSR Scatter Plot}
\usage{
plot_tsr_scatter(experiment, sample_1, sample_2)
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized TSR counts}

\item{sample_1}{First sample name to plot}

\item{sample_2}{Second sample name to plot}
}
\value{
ggplot2 object
}
\description{
Scatter plots to explore replicate concordance.
}
