% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plots.R
\name{plot_tss_average}
\alias{plot_tss_average}
\title{TSS Average Plots}
\usage{
plot_tss_average(experiment, sample, upstream = 1000,
  downstream = 1000, threshold = 1)
}
\arguments{
\item{experiment}{tsr_explorer object with annotated TSSs}

\item{upstream}{Bases upstream of plot center}

\item{downstream}{BZZases downstream of plot center}

\item{threshold}{threshold value for TSSs}
}
\value{
ggplot2 object of average plot
}
\description{
Generate average plot of TSSs versus annotated TSSs
}
