% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\docType{methods}
\name{plot_tss_corr}
\alias{plot_tss_corr}
\alias{plot_tss_corr,tsr_object-method}
\title{Replicate Correlation Matrix}
\usage{
plot_tss_corr(experiment, ...)

\S4method{plot_tss_corr}{tsr_object}(experiment,
  corr_metric = c("pearson", "spearman"))
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized counts}

\item{corr_metric}{Correlation metric ("pearson", "spearman")}
}
\value{
ggplot2 object
}
\description{
Correlation matrix to explore replicate concordance.
}
