% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_tss_heatmap}
\alias{plot_tss_heatmap}
\title{Plot TSS Heatmap}
\usage{
plot_tss_heatmap(tss_heatmap_matrix, max_value = 5)
}
\arguments{
\item{max_value}{Max log2 value to truncate heatmap color}

\item{tss_count_matrix}{TSS count matrix from tss_count_matrix}
}
\value{
ggplot2 object of TSS heatmap
}
\description{
Plot TSS heatmap from count matrix generated by tss_count_matrix
}
