% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\docType{methods}
\name{plot_tss_scatter}
\alias{plot_tss_scatter}
\alias{plot_tss_scatter,tsr_object-method}
\title{Replicate Scatter Plot}
\usage{
plot_tss_scatter(experiment, ...)

\S4method{plot_tss_scatter}{tsr_object}(experiment, sample_1, sample_2)
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized counts}

\item{sample_1}{First sample name to plot}

\item{sample_2}{Second sample name to plot}
}
\value{
ggplot2 object
}
\description{
Scatter plots to explore replicate concordance.
}
