% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_ma}
\alias{plot_ma}
\title{DE MA Plot}
\usage{
plot_ma(differential_expression, data_type = c("tss", "tsr",
  "tss_features", "tsr_features"), de_comparisons = "all", ncol = 1,
  ...)
}
\arguments{
\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{Which differential expression comparisons to plot}

\item{ncol}{Number of columns for the facets}

\item{...}{Arguments passed to geom_point}

\item{experiment}{tsrexplorer object}
}
\value{
ggplot2 object of differential TSRs volcano plot.
}
\description{
Generate MA plot for differential TSRs or Genes (RNA-seq)
}
