% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{DE Volcano Plot}
\usage{
plot_volcano(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  de_comparisons = "all",
  log2fc_cutoff = 1,
  fdr_cutoff = 0.05,
  ncol = 1,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{either 'tss', 'tsr', 'tss_features', or 'tsr_features'.}

\item{de_comparisons}{Character vector of differential expression comparisons to plot.}

\item{log2fc_cutoff}{Differential features not meeting this |Log2 FC| threshold will not be considered.}

\item{fdr_cutoff}{Differential features not meeting this significance threshold will not be considered.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{...}{Arguments passed to geom_point.}
}
\description{
Generate a volcano plot for differential TSSs, TSRs, or genes/transcripts.
}
