% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{preliminary_filter}
\alias{preliminary_filter}
\title{Preliminary Filter}
\usage{
preliminary_filter(signal_data, dominant, threshold)
}
\arguments{
\item{signal_data}{TSS or TSR data}

\item{dominant}{Whether to retain dominant TSS/TSR}

\item{threshold}{Wheher to apply a threshold to the scores}
}
\description{
Preliminary filter of data
}
