% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{ShiftScores}
\alias{ShiftScores}
\title{Shifting Score}
\usage{
ShiftScores(
  tss_table,
  baseline_level,
  calc_pvalue = TRUE,
  nresamp = 100L,
  nthresh = 2,
  check_sort = TRUE
)
}
\arguments{
\item{tss_table}{Table of TSSs perpared for shifting score calculation}

\item{baseline_level}{The sample being used as the baseline for calculation}

\item{calc_pvalue}{Whether p-values should be returned for comparisons}

\item{nresamp}{Number of resamplings for the permutation test}

\item{nthresh}{Both samples must have at least this number of reads in each TSR}

\item{check_sort}{Check that the input is sorted properly (by fhash? qq)}
}
\description{
Calculate shifting scores and associated permutation test p-values.
}
