% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softclip_analysis.R
\name{softclip_histogram}
\alias{softclip_histogram}
\title{Number of Softclipped Bases}
\usage{
softclip_histogram(experiment, samples = "all", n_bases = NULL)
}
\arguments{
\item{experiment}{tsr explorer object}
}
\description{
Number of Softclipped Bases
}
