% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{tsr_annotation}
\alias{tsr_annotation}
\title{Annotate TSRs}
\usage{
tsr_annotation(experiment, annotation_file, feature_type = c("gene",
  "transcript"), upstream = 1000, downstream = 100)
}
\arguments{
\item{experiment}{tsrexplorer object with TSR Granges}

\item{upstream}{Bases upstream of TSR}

\item{downstream}{Bases downstream of TSR}

\item{gtf}{GTF genomic annotation file}
}
\value{
Annotated TSR slot in tsrchitect object
}
\description{
Use ChIPseeker to annotate TSRs
}
