% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{tsr_experiment}
\alias{tsr_experiment}
\alias{tsr_experiment,tsr_explorer-method}
\title{Add TSRs}
\usage{
tsr_experiment(TSRexploreR_object)

\S4method{tsr_experiment}{tsr_explorer}(TSRexploreR_object)
}
\arguments{
\item{TSRexploreR_object}{TSRexploreR object}
}
\description{
Convenience function to add TSRs to a TSRexploreR object.
}
