% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrexplorer.R
\docType{class}
\name{tsr_explorer-class}
\alias{tsr_explorer-class}
\alias{tsr_explorer}
\title{TSRexplorer Class}
\usage{
tsr_explorer(TSSs = NA, TSRs = NA, cores = 1)
}
\arguments{
\item{TSSs}{Named list of TSS GRanges returned by TSRchitect}

\item{TSRs}{Named list of TSR GRanges returned by TSRchitect}

\item{cores}{Number of CPU cores/threads available}
}
\value{
A tsrexplorer object containing TSSs and/or TSRs
}
\description{
This function generates a new tsr_explorer object for
detailed analysis of transcription start sites (TSSs)
and TSS clusters, referred to as transcription
start regions (TSRs) or clustered transcription start sites (cTSSs).
}
\section{Slots}{

\describe{
\item{\code{experiment}}{Named lists containing GRanges of TSSs and/or TSRs}

\item{\code{cores}}{Integer value for number of cores available}

\item{\code{annotated}}{Named lists of annotated TSSs and/or TSRs}

\item{\code{counts}}{Named lists of TMM and CPM normalized TSSs and/or TSRs}

\item{\code{correlation}}{Named lists of correlation values between TSS and/or TSR sets}
}}

\examples{
TSSs <- system.file("extdata", "yeast_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
exp <- tsr_explorer(TSSs)

}
