% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{tsr_export}
\alias{tsr_export}
\title{Export TSRs}
\usage{
tsr_export(
  experiment,
  samples = "all",
  file_type = "bed",
  out_dir = NA,
  diff_tsr = FALSE,
  sep = "\\t"
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{file_type}{Either 'bed' or 'table'.}

\item{out_dir}{Output directory for files.}

\item{diff_tsr}{If TRUE, will output differential TSSs.}

\item{sep}{Delimiter for tabular output.}
}
\value{
Either a BED file or a table.
}
\description{
Export TSRs to table or BED
}
\details{
This function will save TSRs as BED files or a delimited table.

'file_type' controls what the TSRs will be output as. 'bed' will result in each 
sample being saved as a BED file. 'table' will output a file with the delimiter 
specified by the 'sep' argument. The resulting table will have all columns added 
to the TSR data in the TSRexplorer object, such as annotation information.

The directory to output the files can be set with 'out_dir'. A value of NA will 
save the files to the working directory by default.

If 'diff_tsr' is TRUE, only differential TSRs will be output.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)

tsre <- TSSs[1] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3)
\donttest{tsr_export(tsre)}

}
\seealso{
\code{\link{tss_export}} to export TSSs.
\code{\link{tss_import}} to import TSSs.
\code{\link{tsr_import}} to import TSRs.
}
