% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{tsr_genomic_distribution}
\alias{tsr_genomic_distribution}
\title{TSR Genomic Distribution}
\usage{
tsr_genomic_distribution(experiment, sample)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSRs}

\item{sample}{Name of sample to analyze}
}
\value{
tibble with TSR genomic distribution stats
}
\description{
Get genomic distribution of TSRs
}
