% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_metrics.R
\name{tsr_metrics}
\alias{tsr_metrics}
\title{TSR Metrics}
\usage{
tsr_metrics(experiment, iqr_lower = 0.1, iqr_upper = 0.9)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{iqr_lower}{Lower IQR cutoff value.}

\item{iqr_upper}{Upper IQR cutoff value.}
}
\description{
Calculate TSR Metrics.
}
