% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_metrics.R
\name{tsr_metrics}
\alias{tsr_metrics}
\title{TSR Metrics}
\usage{
tsr_metrics(experiment, iqr_lower = 0.1, iqr_upper = 0.9)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{iqr_lower}{Lower IQR cutoff value.}

\item{iqr_upper}{Upper IQR cutoff value.}
}
\value{
TSRexplorer object with TSR metrics added to
  TSSs and TSRs.
}
\description{
Calculate various TSR Metrics including:
  IQR width, peak balance and shape index.

TSR shape descriptors such as peaked versus broad can enrich for
  different functional classes of genes.
TSRexploreR allows for the calculation of various common TSR shape indicators
  such as Interquartile width (IQR), shape index (broad versus peaked), and peak balance.
'iqr_lower' and 'iqr_upper' determine the upper and lower quantile bounds for calculation.
}
\examples{
data(TSSs_reduced)

tsre <- TSSs \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  associate_with_tsr
tsre <- tsr_metrics(tsre)

}
