% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{tsr_normalization}
\alias{tsr_normalization}
\title{TMM Normalize TSRs}
\usage{
tsr_normalization(experiment)
}
\arguments{
\item{experiment}{tsrexplorer object with TSR GRanges}
}
\value{
tibble of TMM normalized read counts
}
\description{
Using edgeR to TMM normalize TSRs
}
