% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\docType{methods}
\name{tsrexplorer_import}
\alias{tsrexplorer_import}
\alias{tsrexplorer_import,character-method}
\alias{tsrexplorer_import,data.frame-method}
\alias{tsrexplorer_import,tssObject-method}
\alias{tsrexplorer_import,CAGEexp-method}
\title{Import TSSs and TSRs}
\usage{
tsrexplorer_import(object, data_type)

\S4method{tsrexplorer_import}{character}(object, data_type)

\S4method{tsrexplorer_import}{data.frame}(object, data_type)

\S4method{tsrexplorer_import}{tssObject}(object, data_type)

\S4method{tsrexplorer_import}{CAGEexp}(object, data_type)
}
\arguments{
\item{object}{Object to import into tsrexplorer}

\item{data_type}{Whether the data being imported is 'TSSs' or 'TSRs'}
}
\description{
Convenience function to import TSSs and/or TSRs from various sources.
}
