% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{tss_aggregate}
\alias{tss_aggregate}
\title{Aggregate TSSs}
\usage{
tss_aggregate(experiment)
}
\arguments{
\item{experiment}{tsr explorer object}
}
\description{
Aggregate overlapping TSSs into a total sum score.
}
