% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{tss_annotation}
\alias{tss_annotation}
\title{Annotate TSSs}
\usage{
tss_annotation(experiment, annotation_file, feature_type = c("gene",
  "transcript"), upstream = 1000, downstream = 100)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS Granges}

\item{upstream}{Bases upstream of TSS}

\item{downstream}{Bases downstream of TSS}

\item{gtf}{GTF genomic annotation file}
}
\value{
Annotated TSS slot in tsrchitect object
}
\description{
Use ChIPseeker package to annotate TSSs.
}
