% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_clustering.R
\name{tss_clustering}
\alias{tss_clustering}
\title{TSS Clustering}
\usage{
tss_clustering(experiment, samples = "all", threshold = 1,
  max_distance = 25)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{The samples to call TSRs for}

\item{threshold}{Minimum number of reads for a TSS to be considered}

\item{max_distance}{The maximum distance to cluster TSSs}
}
\description{
Basic function to cluster TSSs
}
