% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\docType{methods}
\name{tss_experiment}
\alias{tss_experiment}
\alias{tss_experiment,tsr_explorer-method}
\title{Add TSSs}
\usage{
tss_experiment(tsrexplorer_object)

\S4method{tss_experiment}{tsr_explorer}(tsrexplorer_object)
}
\arguments{
\item{tsrexplorer_object}{tsr_explorer object}
}
\description{
Convenience function to add TSSs to tsrexplorer object
}
