% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{tss_export}
\alias{tss_export}
\title{Export TSSs}
\usage{
tss_export(experiment, samples = "all", file_type = "bedgraph",
  out_dir = NA, diff_tss = FALSE)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{Names of samples to export (is this one of those ones where it could also be 'all'?)}

\item{file_type}{either 'bedgraph' or 'table'}

\item{out_dir}{Output directory for files}

\item{diff_tss}{Whether to pull out the differential TSSs (qq????)}
}
\description{
Export TSSs to table or bedgraph
}
