% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{tss_heatmap_matrix}
\alias{tss_heatmap_matrix}
\title{TSS Heatmap Count Matrix}
\usage{
tss_heatmap_matrix(experiment, samples = "all", upstream = 1000,
  downstream = 1000, threshold = NA, use_cpm = FALSE,
  dominant = FALSE, data_conditions = list(order_by = "score"))
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs}

\item{samples}{Either 'all' or name of samples to analyze}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}

\item{threshold}{Reads required per TSS}

\item{use_cpm}{Whether to use the CPM normalized values or not}

\item{dominant}{Consider only dominant}

\item{data_conditions}{Condition the data (filter, order, and quantile/group available)}
}
\value{
DataFrame of counts for each gene/transcript and position
}
\description{
Generate count matrix to make TSS heatmap
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
hm_mat <- tss_heatmap_matrix(tsre_exp)

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
\code{\link{plot_heatmap}} to plot the heatmap.
\code{\link{tsr_heatmap_matrix}} to generate the TSR matrix data for plotting.
}
