% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tss_import}
\alias{tss_import}
\title{Import TSSs}
\usage{
tss_import(experiment, sample_sheet = NULL, file_type = "auto", delim = "\\t")
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{sample_sheet}{Sample sheet}

\item{file_type}{Either 'auto', 'bedgraph', 'bigwig', or 'table'}

\item{delim}{If the input is a table, use this delimiter}
}
\value{
tsr explorer object with imported TSSs
}
\description{
Function to import TSSs from various sources.
}
\details{
TSRexploreR can import TSSs from various sources.
Currently bedgraphs, bigwigs, and TSRchitect tssObjects are supported.

To import bedgraphs or bigwigs, a sample sheet must first be created.
The sample sheet can be a data.frame or tabular file.
It should have three columns: sample_name, file_1, and file_2.
sample_name specifies the sample name that will be added for that TSS in the
  tsr explorer object.
file_1 and file_2 should be the path to the two bedgraphs or bigwigs,
  which are usually separated by strand.

To import TSSs directly from a TSRchitect tssObject,
  the TSRchitect workflow must first be run up to the
  'processTSS' step.
}
\seealso{
\code{\link{tsr_import}} to import TSRs.
\code{\link{tss_export}} to export TSSs.
\code{\link{tsr_export}} to export TSRs.
}
