% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSSs}
\alias{TSSs}
\title{Yeast STRIPE-seq TSSs}
\format{
A list of GRanges:
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE142524}{GSE14254}
}
\usage{
TSSs
}
\description{
3 Untreated and 3 Diamide treated yeast TSSs
as determined by STRIPE-seq.
}
\keyword{datasets}
