% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSSs_reduced}
\alias{TSSs_reduced}
\title{STRIPE-seq TSSs YDR077W (Diamide)}
\format{
A named list of one GRange.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE142524}{GSE14254}
}
\usage{
TSSs_reduced
}
\description{
STRIPE-seq TSSs after diamide treatment for gene YDR077W
}
\keyword{datasets}
