% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_groups.R
\name{add_sample_sheet}
\alias{add_sample_sheet}
\title{Sample Sheet}
\usage{
add_sample_sheet(experiment, sample_sheet)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{sample_sheet}{Sample sheet as tab delimited file or data.frame.
Must have columns: 'tss_name' and 'tsr_name', and can have additional
  columns specifying condition, batch, etc.}
}
\description{
Add a sample sheet to a tsrexplorer object.
}
