% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_features}
\alias{annotate_features}
\title{Annotate Data}
\usage{
annotate_features(experiment, annotation_data, data_type = c("tss",
  "tsr", "tss_diff", "tsr_diff"), feature_type = c("gene", "transcript"),
  upstream = 1000, downstream = 100)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS Granges}

\item{annotation_data}{Either path and file name of annotation file, or TxDb object of annotation}

\item{data_type}{Whether to annotate TSSs or TSRs}

\item{feature_type}{Annotate on gene or transcript level}

\item{upstream}{Bases upstream of TSS}

\item{downstream}{Bases downstream of TSS}
}
\value{
tsrexplorer object with annotated features
}
\description{
Use ChIPseeker package to annotate TSSs or TSRs.
}
\details{
This function attempts to assign TSSs or TSRs to the nearest genomic feature.
Genomic annotation data can be provided as either a 'GTF' or 'GFF' file,
  or as a TxDb package from bioconductor.

'feature_type' will let you assign as gene or transcript level.
Furthermore, the promoter region can be defined using
  'upstream' and 'downstream' which are relative to the TSSs
  defined in your annotation data.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)

}
