% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply a threshold to TSSs or TSRs}
\usage{
apply_threshold(experiment, threshold, n_samples = 1, use_normalized = FALSE)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{n_samples}{Number of samples threshold must be reached to keep TSS.
By default set to 1 sample. A NULL value will result in all samples being
required to have read counts above threshold.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}
}
\description{
Apply a threshold to TSSs or TSRs
}
