% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply a threshold to TSSs or TSRs}
\usage{
apply_threshold(experiment, threshold, n_samples = 1, use_normalized = FALSE)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{n_samples}{Number of samples threshold must be reached to keep TSS.
By default set to 1 sample. A NULL value will result in all samples being
required to have read counts above threshold.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}
}
\value{
TSRexploreR object with weak TSSs filtered out of counts table.
}
\description{
Filter TSSs based on given threshold.
}
\details{
All TSSs mapping methods produce spurious 5' reads.
For the most part, these spurious reads tend to be weak and somewhat uniformly
  distributed through the promoter and gene body.
This means that this background can be mitigated by requiring a minimum read threshold
  for a TSS to be retained in the filtered dataset.

This function will permantly filter TSSs from the TSS data.table if no sample has
 at least 'threshold' number of reads in at least 'n_samples' number of samples.
}
\examples{
data(TSSs)
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")

tsre <- TSSs[1] \%>\%
  tsr_explorer(genome_annotation=annotation) \%>\%
  format_counts(data_type="tss")
apply_threshold(tsre, threshold=3)

}
\seealso{
\code{\link{plot_threshold_exploration}} to explore fraction of
  promoter proximal TSSs, and absolute number of detected genes.
}
