% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{associate_with_tsr}
\alias{associate_with_tsr}
\title{Associate TSSs}
\usage{
associate_with_tsr(experiment, sample_list = NA,
  use_sample_sheet = FALSE)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{sample_list}{If 'use_sample_sheet' is FALSE, provide list with TSR as name and vector of TSSs to associate with it.}

\item{use_sample_sheet}{Whether to associate TSSs with their matching TSR in the sample sheet}
}
\description{
Associate TSSs with TSRs
}
\details{
tsr explorer provides many options for the import and merging of TSSs and TSRs.
Because of this, TSS samples must be associated with TSR samples after TSR import,
  TSR merging, or TSS clustering using this function.
This adds an extra workflow step, but provides more flexibility for analysis of results.
Each TSS within the sample will be linked to the TSR of the specified sample
  with overlapping genomic coordinates.
TSSs without an overlapping TSR will not be associated to any TSR.

TSS samples can be associated with TSR samples using a list or sample sheet format.
'sample_list' should be a named list of character vectors, with the names being the TSR
  sample name, and the character vector as the TSS samples(s) that should be associated with
  that TSR sample.
If 'use_sample_sheet' is set to true, a previously added sample sheet will be used to group
  TSR samples with TSS samples.
The sample sheet can be added to the tsr explorer object using the 'add_sample_sheet' function.
It should have 3 columns: 'replicate_id', 'tss_name', and 'tsr_name'.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
tsre_exp <- tss_clustering(tsre_exp)
tsre_exp <- associate_with_tsr(tsre_exp, sample_list = list(
  "S288C_WT_1" = "S288C_WT_1", "S288C_WT_2" = "S288C_WT_2", "S288C_WT_3" = "S288C_WT_3",
  "S288C_D_1" = "S288C_D_1", "S288C_D_2" = "S288C_D_2", "S288C_D_3" = "S288C_D_3"
))

}
