% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{associate_with_tsr}
\alias{associate_with_tsr}
\title{Associate TSSs}
\usage{
associate_with_tsr(experiment, sample_list = NA, use_sample_sheet = FALSE)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{sample_list}{If 'use_sample_sheet' is FALSE, provide a list with TSR and TSS sample names}

\item{use_sample_sheet}{Whether to use a sample sheet as a key for association of TSS and TSR samples}
}
\description{
Associate TSSs with TSRs
}
\details{
TSRexploreR provides many options for the import and merging of TSSs and TSRs.
Because of this, TSS samples must be associated with TSR samples after TSR import,
  TSR merging, or TSS clustering using this function.
This adds an extra workflow step, but provides more analytical flexibility (qq meaning what?).
Each TSS with genomic coordinates overlapping those of a TSR in the specified TSR sample
  will be linked to that TSR.
TSSs not overlapping a TSR in the specified sample will not be associated with any TSR.

TSS samples can be associated with TSR samples using a list or sample sheet.
'sample_list' should be a named list of character vectors, with the names being the TSR
  sample names and the character vectors as the TSS samples(s) that should be associated with
  each TSR sample.
If 'use_sample_sheet' is set to true, a previously added sample sheet will be used to associate
  TSR samples with TSS samples.
The sample sheet can be added to the TSRexploreR object using the 'add_sample_sheet' function.
It should have 3 columns: 'replicate_id', 'tss_name', and 'tsr_name'.

#' @seealso
\code{\link{add_sample_sheet}} to add a sample sheet to the TSRexploreR object.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- tss_clustering(tsre_exp)
tsre_exp <- associate_with_tsr(tsre_exp, sample_list=list(
  "S288C_WT_1"="S288C_WT_1", "S288C_WT_2"="S288C_WT_2", "S288C_WT_3"="S288C_WT_3",
  "S288C_D_1"="S288C_D_1", "S288C_D_2"="S288C_D_2", "S288C_D_3"="S288C_D_3"
))

}
