% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{common_params}
\alias{common_params}
\title{Function to hold common parameters.}
\usage{
common_params(x)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{genome_assembly}{Genome assembly in FASTA or BSgenome format.}

\item{genome_annotation}{Genome annotation in GTF, GFF3, or TxDb format.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{samples}{A vector of sample names to analyze.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{log2fc_cutoff}{Differential features not meeting this |Log2 FC| threshold will not be considered.}

\item{fdr_cutoff}{Differential features not meeting this significance threshold will not be considered.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{exclude_antisense}{Remove antisense TSSs/TSRs prior to analysis.}

\item{data_conditions}{Apply advanced conditions to the data.}

\item{rasterize}{Rasterize a ggplot.}

\item{raster_dpi}{If rasterization is set, this controls the rasterization DPI.}
}
\description{
Function to hold common parameters.
}
