% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{count_matrix}
\alias{count_matrix}
\title{Count Matrix}
\usage{
count_matrix(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"), samples = "all")
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{samples}{The samples to turn into a count matrix}
}
\description{
Generate count matrices
}
