% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{count_normalization}
\alias{count_normalization}
\title{TMM Normalize TSSs or TSRs}
\usage{
count_normalization(experiment, data_type = c("tss", "tsr"))
}
\arguments{
\item{experiment}{tsrexplorer object}
}
\value{
tibble of TMM normalized read counts
}
\description{
Using edgeR to TMM normalize TSSs or TSRs
}
