% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{cpm_normalize}
\alias{cpm_normalize}
\title{CPM Normalize Counts}
\usage{
cpm_normalize(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"))
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{'tss', 'tsr', 'tss_features', 'tsr_features'}
}
\description{
CPM normalize the TSS, TSR, and/or feature counts.
}
