% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{cpm_normalize}
\alias{cpm_normalize}
\title{CPM Normalize Counts}
\usage{
cpm_normalize(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features")
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{'tss', 'tsr', 'tss_features', or 'tsr_features'}
}
\value{
TSRexploreR object with CPM values.
}
\description{
CPM normalize the TSS, TSR, and/or feature counts.
}
\details{
Counts Per Million (CPM) is an approach for read number normalization, 
  allowing comparison between samples for various downstream analyses and plots.
For more quantitative comparisons, TMM normalization is recommended (this should be fleshed out more qq)
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- cpm_normalize(tsre_exp, data_type="tss")

}
