% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{detect_features}
\alias{detect_features}
\title{Genes/Transcripts Detected}
\usage{
detect_features(experiment, samples = "all", data_type = c("tss",
  "tsr"), threshold = NA, dominant = FALSE, condition_data = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs or TSRs}

\item{samples}{Either 'all' or vector of sample names}

\item{data_type}{Whether TSSs or TSRs should be analyzed}

\item{threshold}{The number of reads required in a TSS or TSR to avoid filtering}

\item{dominant}{Whether to consider only the dominant TSS or TSR}

\item{condition_data}{Apply conditions to data (supports filtering and quantiles/grouping)}
}
\value{
DataFrame of detected feature numbers.
}
\description{
Get the number of genes or transcripts detected
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
detected <- detect_features(tsre_exp, data_type = "tss")

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs and TSRs.
\code{\link{plot_detected_features}} to plot detected features.
}
