% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{detect_features}
\alias{detect_features}
\title{Genes/Transcripts Detected}
\usage{
detect_features(experiment, samples = "all", data_type = c("tss",
  "tsr"), threshold = NA, dominant = FALSE, condition_data = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs or TSRs}

\item{samples}{Either 'all' or a vector of sample names}

\item{data_type}{Whether TSSs or TSRs should be analyzed}

\item{threshold}{The number of raw reads required for a TSS or TSR to be considered}

\item{dominant}{Whether to consider only the dominant TSS or TSR}

\item{condition_data}{Apply conditions to data (supports filtering and quantiles/grouping)}
}
\value{
DataFrame of detected feature numbers.
}
\description{
Get the number of genes or transcripts with an associated unique TSS or TSR
}
\details{
This function will return either the number of genes or transcripts with
 an associated unique TSS or TSR.
Information on whether the feature has a promoter-proximal TSS or TSR is included
  in the output for plotting purposes.

A set of functions to control data structure for plotting are included.
'threshold' will define the minimum number of reads a TSS or TSR
 must have to be considered.
'dominant' specifies whether only the dominant TSS or TSR is considered 
  from the 'mark_dominant' function.
For TSSs this can be either dominant per TSR or gene, and for TSRs
  it is just the dominant TSR per gene.
'data_conditions' allows for the advanced filtering, ordering, and grouping
  of data.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
detected <- detect_features(tsre_exp, data_type = "tss")

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs and TSRs.
\code{\link{plot_detected_features}} to plot detected features.
}
