% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Find Differential TSRs}
\usage{
differential_expression(fit_edger_model, data_type = c("tsr", "rnaseq"),
  compare_groups = c())
}
\arguments{
\item{fit_edger_model}{edgeR differential expression model from fit_edger_model}

\item{data_type}{Whether the input was made from TSRs or RNA-seq}

\item{compare_groups}{Vector of length two of the two groups to find differential TSRs}
}
\value{
tibble of differential TSRs
}
\description{
Find differential TSRs from edgeR model.
}
