% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Find Differential Expression}
\usage{
differential_expression(experiment, data_type = c("tss", "tsr",
  "tss_features", "tsr_features"), compare_groups, fdr_cutoff = 0.05,
  log2fc_cutoff = 1)
}
\arguments{
\item{data_type}{Whether the input was made from TSSs, TSRs, or features}

\item{compare_groups}{Vector of length two of the two groups to find differential TSRs}

\item{fit_edger_model}{edgeR differential expression model from fit_edger_model}
}
\value{
tibble of differential TSRs
}
\description{
Find differential TSSs, TSRs, or features from edgeR model.
}
