% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Analyze Differential Expression}
\usage{
differential_expression(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  comparison_name,
  comparison_type,
  comparison,
  fdr_cutoff = 0.05,
  log2fc_cutoff = 1,
  shrink_lfc = FALSE
)
}
\arguments{
\item{experiment}{tsrexplorer object with edgeR differential expression model from fit_edger_model}

\item{data_type}{Whether the input was generated from TSSs, TSRs, or features}

\item{comparison_name}{The name given to the comparison when stored back into the tsr explore robject.}

\item{comparison_type}{For DEseq2 either 'contrast' or 'name'.
For edgeR either 'contrast' or 'coef'.}

\item{comparison}{For edgeR either the coefficients or contrasts.
For DESeq2 the contrast or name.}

\item{fdr_cutoff}{FDR cutoff}

\item{log2fc_cutoff}{Log2 fold change cutoff}

\item{shrink_lfc}{For DESeq2 whether the Log2 Fold Changes are shrunk (TRUE) or left alone (FALSE).}
}
\value{
tibble of differential TSRs
}
\description{
Find differential TSSs, TSRs, or features from edgeR model
}
