% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Analyze Differential Expression}
\usage{
differential_expression(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  compare_groups,
  fdr_cutoff = 0.05,
  log2fc_cutoff = 1
)
}
\arguments{
\item{experiment}{TSRexploreR object with edgeR differential expression model from fit_edger_model}

\item{data_type}{Whether the input was generated from TSSs, TSRs, or features}

\item{compare_groups}{Vector of length two of the two groups from which to find differential TSRs}

\item{fdr_cutoff}{FDR cutoff}

\item{log2fc_cutoff}{Log2 fold change cutoff}
}
\value{
tibble of differential TSRs
}
\description{
Find differential TSSs, TSRs, or features from edgeR model
}
