% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotide.R
\name{dinucleotide_frequencies}
\alias{dinucleotide_frequencies}
\title{Dinucleotide Analysis}
\usage{
dinucleotide_frequencies(experiment, genome_assembly, samples = "all",
  threshold = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS GRanges}

\item{genome_assembly}{fasta file of genome assembly}

\item{samples}{Either 'all' or vector of sample names to analyze}

\item{threshold}{TSS read threshold}
}
\value{
tibble with dinucleotide frequencies
}
\description{
Analysis of -1 and +1 dinucleotide frequencies.
}
