% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.annotate}
\alias{.annotate}
\title{Annotate Features}
\usage{
.annotate(sample_table, upstream, downstream, feature_type, annotation_data)
}
\arguments{
\item{sample_table}{Sample table}

\item{upstream}{Number of bases upstream of TSS for 'promoter' annotation (integer).}

\item{downstream}{Number of bases downstream of TSS for 'promoter' annotation (integer).}

\item{feature_type}{Whether to annotate at the 'gene' or 'transcript' level.}

\item{annotation_data}{Genome annotation.}
}
\description{
Annotate Features
}
