% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{.calculate_distribution}
\alias{.calculate_distribution}
\title{Calculate Genomic Distribution}
\usage{
.calculate_distribution(selected_samples, groupings)
}
\arguments{
\item{selected_samples}{Samples to analyze}

\item{groupings}{Whether data is grouped}
}
\description{
Calculate Genomic Distribution
}
