% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{.count_features}
\alias{.count_features}
\title{Calculate Feature Counts}
\usage{
.count_features(sample_data, data_type, groupings)
}
\arguments{
\item{sample_data}{Sample data}

\item{data_type}{Type of data}

\item{groupings}{Whether there is a grouping variable}
}
\description{
Calculate Feature Counts
}
