% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_matrix.R
\name{.count_matrix}
\alias{.count_matrix}
\title{Generate Count Matrix.}
\usage{
.count_matrix(count_data, data_type = c("tss", "tsr"), use_normalized = FALSE)
}
\arguments{
\item{count_data}{Names list of counts.}

\item{data_type}{Currently supports 'tss' only.}

\item{use_normalized}{Whether to use normalized counts.}
}
\description{
Generate Count Matrix.
}
