% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.de_status}
\alias{.de_status}
\title{Mark DE Status}
\usage{
.de_status(de_results, log2fc_cutoff, fdr_cutoff)
}
\arguments{
\item{de_results}{Results of DE}

\item{log2fc_cutoff}{Values below the |Log2 FC| cutoff will be discarded.}

\item{fdr_cutoff}{Values below the FDR cutoff will be discarded.}
}
\description{
Mark DE Status
}
