% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.de_status}
\alias{.de_status}
\title{Mark DE Status}
\usage{
.de_status(de_results, log2fc_cutoff, fdr_cutoff)
}
\arguments{
\item{de_results}{Results of DE.}

\item{log2fc_cutoff}{Differential features not meeting this |log2(fold change)| threshold will not be considered.}

\item{fdr_cutoff}{Differential features not meeting this significance threshold will not be considered.}
}
\description{
Mark DE Status
}
