% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.deseq2_model}
\alias{.deseq2_model}
\title{DESeq2 Differential Expression Model}
\usage{
.deseq2_model(count_data, sample_sheet, formula)
}
\arguments{
\item{count_data}{Count matrix}

\item{sample_sheet}{Sample data}

\item{formula}{Differential expression formula}
}
\description{
DESeq2 Differential Expression Model
}
