% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{.edger_model}
\alias{.edger_model}
\title{edgeR Differential Expression Model}
\usage{
.edger_model(count_data, sample_sheet, formula)
}
\arguments{
\item{count_data}{Count matrix.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{formula}{Differential expression formula.}
}
\description{
edgeR Differential Expression Model
}
