% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{explore_thresholds}
\alias{explore_thresholds}
\title{Threshold Exploration}
\usage{
explore_thresholds(experiment, max_threshold = 50, samples = "all")
}
\arguments{
\item{experiment}{tsrexplorer object with TSSs}

\item{max_threshold}{Threshold from 1 to max_threshold will be explored}

\item{samples}{Either 'all' or the names of the samples to analyze}
}
\value{
data.frame containing information for each threshold and sample
}
\description{
Explore various naive thresholds.
}
\details{
All TSS mapping technologies have some degree of spurious background reads.
tsr_explorer allows for basic thresholding of these background reads by requiring
  a minimum number of reads for a TSS to be considered.

An easy way to pick a threshold is with a threshold plot.
This plot shows the proportion of promoter-proximal TSS positions when
  various naive threshold values are tested.
'max_threshold' defines the maximum threshold value tested from 1 to that value.
A recommended final value will likely result in a promoter-proximal fraction above .80 to .85
  and without sacrificing the ability to detect too many unique genes or transcripts.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
thresh_results <- explore_thresholds(tsre_exp)

}
\seealso{
\code{\link{plot_threshold_exploration}} to plot the results.
}
