% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{export_for_enrichment}
\alias{export_for_enrichment}
\title{Export to clusterProfiler}
\usage{
export_for_enrichment(annotated_differential_tsrs, log2fc_cutoff = 1,
  fdr_cutoff = 0.05)
}
\arguments{
\item{annotated_differential_tsrs}{Annotated differential TSRs}

\item{log2fc_cutoff}{Log2 fold change cutoff for significance}

\item{fdr_cutoff}{FDR cutoff for significance}
}
\description{
Export DEGs for use in clusterProfiler term enrichment.
}
