% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{export_for_enrichment}
\alias{export_for_enrichment}
\title{Export to clusterProfiler}
\usage{
export_for_enrichment(
  experiment,
  data_type = c("tss", "tsr"),
  de_comparisons = "all",
  log2fc_cutoff = 1,
  fdr_cutoff = 0.05,
  keep_unchanged = FALSE
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The DE comparisons to plot}

\item{log2fc_cutoff}{Log2FC cutoff value}

\item{fdr_cutoff}{FDR cutoff value}

\item{keep_unchanged}{If TRUE genes that are labelled unchanged will
be kept.}
}
\description{
Export DEGs for use in clusterProfiler term enrichment.
}
