% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{export_for_enrichment}
\alias{export_for_enrichment}
\title{Export to clusterProfiler}
\usage{
export_for_enrichment(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  de_comparisons = "all"
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The DE comparisons to plot}
}
\description{
Export DEGs for use in clusterProfiler term enrichment.
}
