% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract Counts}
\usage{
extract_counts(experiment, data_type, samples, use_normalized = FALSE)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to extract TSS or TSR counts}

\item{samples}{A vector of sample names to analyze.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}
}
\description{
Extract counts from a TSRexploreR object.
}
