% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract Counts}
\usage{
extract_counts(experiment, data_type, samples, use_normalized = FALSE)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Whether to extract TSS or TSR counts}

\item{samples}{Names of samples from which to extract counts}

\item{use_normalized}{Whether CPM values should be returned}
}
\description{
Extract counts from a tsrexplorer object.
}
