% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract Counts}
\usage{
extract_counts(experiment, data_type, samples, use_cpm = FALSE)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{whether to extract from the 'tss' or 'tsr' sets}

\item{samples}{names of samples to extract}

\item{use_cpm}{Whether CPM values should be returned}
}
\description{
Extract counts from a tsrexplorer object.
}
