% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Extract Count Matrices}
\usage{
extract_matrix(experiment, data_type, samples)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Whether to extract 'tss', 'tsr', or 'feature' counts}

\item{samples}{Sampels to extract}
}
\description{
Extract the normalized count matrices.
}
