% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{fit_de_model}
\alias{fit_de_model}
\title{edgeR Model for DE}
\usage{
fit_de_model(
  experiment,
  formula,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  samples = "all",
  method = "DESeq2"
)
}
\arguments{
\item{experiment}{tsrexplorer object with TMM-normalized counts}

\item{formula}{DE formula}

\item{data_type}{Whether TSSs, TSRs, or feature counts should be analyzed}

\item{samples}{Vector of sample names to analyze}

\item{method}{Either 'DESeq2' or 'edgeR'}
}
\value{
DGEList object with fitted model
}
\description{
Find differential TSSs, TSRs, or features
}
