% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{fit_edger_model}
\alias{fit_edger_model}
\title{edgeR Model for DE}
\usage{
fit_edger_model(experiment, data_type = c("tsr", "rnaseq"),
  samples = c(), groups = c())
}
\arguments{
\item{experiment}{tsrexplorer object after TMM normalization.}

\item{samples}{Vector of sample names to analyze.}

\item{groups}{Vector of groups in correct factor notation.}
}
\value{
DGEList object with fitted model
}
\description{
Find differential TSRs (TSS mapping) or Genes (RNAseq)
}
