% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{fit_edger_model}
\alias{fit_edger_model}
\title{edgeR Model for DE}
\usage{
fit_edger_model(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"), samples, groups)
}
\arguments{
\item{experiment}{tsrexplorer object after TMM normalization.}

\item{data_type}{Whether TSSs, TSRs, or feature counts should be analyzed.}

\item{samples}{Vector of sample names to analyze.}

\item{groups}{Vector of groups in correct factor notation.}
}
\value{
DGEList object with fitted model
}
\description{
Find differential TSSs, TSRs, or features
}
