% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{format_counts}
\alias{format_counts}
\title{Format Counts}
\usage{
format_counts(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{'tss' or 'tsr'}

\item{samples}{character vector of sample names}
}
\value{
tsrexplorer object with properly formatted features
   in data.table format.
}
\description{
Format counts for TSSs or TSRs for further analysis.
}
\details{
When TSSs or TSRs are first loaded into the tsrexplorer object
  they are stored as GRanges objects.
This function converts these into data.table format,
  and adds a few important columns for downstream analysis.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")

}
