% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_counts.R
\name{format_counts}
\alias{format_counts}
\title{Format Counts}
\usage{
format_counts(experiment, data_type = c("tss", "tsr"))
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{'tss' or 'tsr'}
}
\description{
Format counts for TSSs, TSRs, and/or features.
}
