% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{G_correction}
\alias{G_correction}
\title{Correct G artifact}
\usage{
G_correction(experiment, assembly = NULL)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{assembly}{Genome assembly in FASTA or BSgenome format.}
}
\description{
Correct overrepresentation of 5' G bases added during reverse transcription.
}
