% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tracks.R
\name{gene_tracks}
\alias{gene_tracks}
\title{Gene Tracks by Gene}
\usage{
gene_tracks(experiment, genome_annotation, feature_name,
  feature_type = "gene", samples = "all", threshold = 1,
  upstream = 250, downstream = 250, promoter_only = FALSE,
  use_cpm = FALSE, tss_colors = "black", tsr_colors = "black",
  axis_scale = 0.25, ymax = NA)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{genome_annotation}{Genome annotation GTF/GFF file, or TxDb object}

\item{feature_name}{Name of gene or transcript to plot}

\item{feature_type}{Either 'gene' or 'transcript'}

\item{samples}{Names of samples to plot.
Append sample names with 'TSS:' or 'TSR:' for TSS and TSR tracks respectively.}

\item{threshold}{TSSs and TSRs below threshold are excluded from plotting}

\item{upstream}{bases upstream to extend gene or promoter track}

\item{downstream}{bases downstream to extend gene or promoter track}

\item{promoter_only}{Instead of plotting the entire gene, plot the promoter region}

\item{use_cpm}{Use CPM normalized reads or not}

\item{tss_colors}{Either a single color value for all TSS tracks, or a vector of colors}

\item{tsr_colors}{Either a single color value for all TSR tracks, or a vector of colors}

\item{axis_scale}{Relative size scale for axis text and title}

\item{ymax}{Maximum value on Y axis for all TSS tracks}
}
\description{
Generate gene tracks in GViz by gene name.
}
