% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Genomic Distribution}
\usage{
genomic_distribution(experiment, sample, data_type = c("tss", "tsr"),
  threshold = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSRs}

\item{sample}{Name of sample to analyze}

\item{data_type}{Whether to get distribution of TSSs or TSRs}

\item{threshold}{Filter out TSSs or TSRs under a certain read count number}
}
\value{
tibble with TSS or TSR genomic distribution stats
}
\description{
Get genomic distribution of TSSs and TSRs
}
