% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Genomic Distribution}
\usage{
genomic_distribution(experiment, data_type = c("tss", "tsr"),
  samples = "all", threshold = NA, dominant = FALSE,
  data_conditions = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSRs}

\item{data_type}{Whether to get distribution of TSSs or TSRs}

\item{samples}{Either "all" or vector of sample names}

\item{threshold}{Filter out TSSs or TSRs under a certain read count number}

\item{dominant}{Should dominant TSS per 'gene' or 'tsr' only be considered}

\item{data_conditions}{List of group options (filtering and quantiles/grouping available)}
}
\value{
DataFrame with TSS or TSR genomic distribution stats
}
\description{
Get genomic distribution of TSSs and TSRs
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
genomic_dist <- genomic_distribution(tsre_exp, data_type = "tss")

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
  \code{\link{plot_genomic_distribution}} to plot the genomic distribution.
}
