% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Genomic Distribution}
\usage{
genomic_distribution(
  experiment,
  data_type = c("tss", "tsr", "shift"),
  samples = "all",
  threshold = NULL,
  use_normalized = FALSE,
  dominant = FALSE,
  data_conditions = NA
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to get distribution of TSSs or TSRs.}

\item{samples}{A vector of sample names to analyze.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{data_conditions}{Apply advanced conditions to the data.}
}
\value{
DataFrame with TSS or TSR genomic distribution stats.
}
\description{
Get genomic distribution of TSSs and TSRs.
}
\details{
This function summarizes the distribution of TSSs or TSRs relative to annotated 
genomic features (exons, introns, intergenic, downstream, and promoter regions).
The promoter region is user-defined during annotation.

A set of functions to control data structure for plotting are included. 'use_normalized' 
will use  normalized scores, which only matters if 'consider_score' is TRUE.
'threshold' defines the minimum number of raw counts a TSS or TSR must have to be 
considered. dominant' specifies whether only the dominant TSS or TSR (determined
using the 'mark_dominant' function) is considered. For TSSs, this can be either 
dominant TSS per TSR or gene/transcript, and for TSRs it is the dominant TSR 
per gene/transcript. 'data_conditionals' can be used to filter, quantile, order, 
and/or group data for plotting.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
exp <- tsr_explorer(TSSs)
exp <- format_counts(exp, data_type="tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")
exp <- annotate_features(
  exp, annotation_data=annotation,
  data_type="tss", feature_type="transcript"
)
genomic_dist <- genomic_distribution(exp, data_type="tss")

}
\seealso{
\code{\link{annotate_features}} to annotate TSSs or TSRs.
  \code{\link{plot_genomic_distribution}} to plot the genomic distribution.
}
