% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_assembly}
\alias{get_assembly}
\title{Get Genome Assembly}
\usage{
get_assembly(experiment)
}
\arguments{
\item{experiment}{TSRexploreR object.}
}
\value{
Either BSgenome or FaFile depending on original source of assembly.
}
\description{
Extract the genome assembly.
}
\examples{
assembly <- system.file("extdata", "S288C_Assembly.fasta", package="TSRexploreR")
tsre <- tsr_explorer(genome_assembly=assembly)
get_assembly(tsre)

}
\seealso{
\code{\link{get_annotation}} to get genome annotation.
}
