% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_counts}
\alias{get_counts}
\title{Get TSS or TSR count tables.}
\usage{
get_counts(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{samples}{A vector of sample names to analyze.}
}
\value{
List of TSS or TSR count data.frame.
}
\description{
Extract the TSS or TSR count tables from the 'counts' slot.
}
\examples{
data(TSSs)

tsre <- tsr_explorer(TSSs[1]) \%>\%
  format_counts(data_type="tss")
tsre <- get_counts(tsre)

}
\seealso{
\code{\link{get_granges}} to get TSS or TSR GRanges.
}
