% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_diff_model}
\alias{get_diff_model}
\title{Get Differential TSS/TSR models}
\usage{
get_diff_model(experiment, data_type = c("tss", "tsr"))
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}
}
\value{
DESeq2 or edgeR differential expression model.
}
\description{
Get the DESeq2 or edgeR model.
}
\examples{
data(TSSs)
sample_sheet <- data.frame(
  sample_name=c(
    sprintf("S288C_D_\%s", seq_len(3)),
    sprintf("S288C_WT_\%s", seq_len(3))
  ),
  file_1=rep(NA, 6), file_2=rep(NA, 6),
  condition=c(
    rep("Diamide", 3),
    rep("Untreated", 3)
  )
)

tsre <- TSSs \%>\%
  tsr_explorer(sample_sheet=sample_sheet) \%>\%
  format_counts(data_type="tss") \%>\%
  fit_de_model(~condition, data_type="tss", method="edgeR")
tsre <- get_diff_model(tsre, data_type="tss")

}
