% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_granges}
\alias{get_granges}
\title{Get TSS or TSR GRanges.}
\usage{
get_granges(experiment, data_type = c("tss", "tsr"), samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss' or 'tsr'.}

\item{samples}{A vector of sample names to analyze.}
}
\value{
List of TSS or TSR GRanges.
}
\description{
Extract the TSS or TSR GRanges from the 'experiment' slot.
}
\examples{
data(TSSs)

tsre <- tsr_explorer(TSSs[1])
tsre <- get_granges(tsre)

}
\seealso{
\code{\link{get_counts}} to get TSS or TSR counts data.frame.
}
