% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_shifting_results}
\alias{get_shifting_results}
\title{Get Shifting Results}
\usage{
get_shifting_results(experiment, samples = "all")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}
}
\value{
List of TSS cluster shifting data.frames.
}
\description{
Extract the TSS cluster shifting results.
}
\examples{
library("magrittr")
data(TSSs)

tsre <- TSSs[c(1, 4)] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  tss_shift(
    sample_1=c(TSS="S288C_WT_1", TSR="S288C_WT_1"),
    sample_2=c(TSS="S288C_D_1", TSR="S288C_D_1"),
    comparison_name="Untreated_vs_Diamide",
    max_distance = 100, min_threshold = 10, n_resamples = 1000L
  )
\donttest{get_shifting_results(tsre)}

}
