% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_data}
\alias{group_data}
\title{Group Data}
\usage{
group_data(signal_data, filters = NA, order_by = NA,
  order_direction = "descending", order_samples = NA,
  order_group = NA, quantile_by = NA, n_quantiles = NA,
  quantile_samples = NA, quantile_group = NA,
  quantile_direction = "descending", grouping = NA)
}
\arguments{
\item{signal_data}{TSS or TSR data}

\item{filters}{Logical string to subset/filter data by}

\item{order_by}{Metric to order data by}

\item{order_direction}{Whether the values should be ordered in
'ascending' or 'descending' order}

\item{quantile_by}{Continuous metric to calculate quantiles}

\item{n_quantiles}{Number of quantiles to calculate for continuous metric}

\item{quantile_samples}{Samples to use when setting quantiles}

\item{quantile_group}{Group these features and quantile based on aggregate mean of feature}

\item{grouping}{If quantiles not set split data by categorical variable}

\item{order_on}{Names of samples to order by}
}
\description{
Group data for plotting
}
