% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_imports.R
\name{import_bams}
\alias{import_bams}
\title{Import BAMs}
\usage{
import_bams(
  experiment,
  paired,
  sample_sheet = NULL,
  soft_remove = 3,
  proper_pair = NULL,
  remove_secondary = TRUE
)
}
\arguments{
\item{experiment}{tsr explorer object}

\item{paired}{Whether the BAMs are paired (TRUE) or unpaired (FALSE)}

\item{soft_remove}{Remove read if greater than this number of soft-clipped bases.}

\item{proper_pair}{Whether reads should be properly paired for paired end data.
TRUE by default when data is paired end.}

\item{sample}{sheet to import bams}

\item{remove_seconday}{Remove non-primary reads.}
}
\description{
Import BAMs
}
