% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{import_bams}
\alias{import_bams}
\title{Import BAMs}
\usage{
import_bams(
  experiment,
  paired,
  sample_sheet = NULL,
  soft_remove = 3,
  proper_pair = NULL,
  remove_secondary = TRUE,
  remove_duplicate = FALSE
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{paired}{Whether the BAMs are paired (TRUE) or unpaired (FALSE).}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{soft_remove}{Remove read if greater than this number of soft-clipped bases 
are present at its 5' end.}

\item{proper_pair}{Whether reads should be properly paired for paired end data.
TRUE by default when data is paired-end.}

\item{remove_duplicate}{If the duplciate flag is set, duplicates are removed.}

\item{remove_seconday}{Remove non-primary reads.}
}
\description{
Import alignment BAM files.
}
