% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge Samples}
\usage{
merge_samples(
  experiment,
  data_type = c("tss", "tsr"),
  threshold = NA,
  merge_replicates = FALSE,
  sample_list = NA
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Whether to merge TSSs or TSRs}

\item{threshold}{Filter out TSSs or TSRs below this raw count threshold before merging}

\item{merge_replicates}{If 'TRUE', replicate groups will be merged}

\item{sample_list}{If merge_replicates is set to 'FALSE',
specify what samples to merge in list format.}
}
\description{
Merge replicates or selected samples.
}
