% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_plots.R
\name{plot_average}
\alias{plot_average}
\title{Average Plots}
\usage{
plot_average(experiment, sample, data_type = c("tss", "tsr"),
  consider_score = FALSE, upstream = 1000, downstream = 1000,
  threshold = 1)
}
\arguments{
\item{experiment}{tsr_explorer object with annotated TSSs}

\item{sample}{Name of sample to make the average plot for}

\item{data_type}{Make average plot for TSS or TSR data}

\item{consider_score}{Should the TSS or TSR score be considered, or just the unique location.}

\item{upstream}{Bases upstream of plot center}

\item{downstream}{Bases downstream of plot center}

\item{threshold}{threshold value for TSSs}
}
\value{
ggplot2 object of average plot
}
\description{
Generate average plots of TSSs or TSRs
}
