% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Replicate TSS Correlation Matrix}
\usage{
plot_correlation(experiment, data_type = c("tss", "tsr", "rnaseq"),
  corr_metric = c("pearson", "spearman"))
}
\arguments{
\item{experiment}{tsrexplorer object with TMM normalized counts}

\item{data_type}{Whether to create a correlation matrix for TSSs, TSRs, or RNA-seq}

\item{corr_metric}{Correlation metric ("pearson", "spearman")}
}
\value{
ggplot2 object
}
\description{
Correlation matrix to explore replicate concordance.
}
