% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot Sample Correlation}
\usage{
plot_correlation(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  samples = "all",
  correlation_metric = "pearson",
  threshold = NULL,
  use_normalized = TRUE,
  font_size = 12,
  cluster_samples = FALSE,
  heatmap_colors = NULL,
  show_values = TRUE,
  return_matrix = FALSE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Whether to make scatter plots from TSS, TSR, or RNA-seq & 5' data.}

\item{samples}{A vector of sample names to analyze.}

\item{correlation_metric}{Whether to use Spearman or Pearson correlation.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{font_size}{The font size for the heatmap tiles.}

\item{cluster_samples}{Logical for whether hierarchical clustering should be performed
on rows and columns.}

\item{heatmap_colors}{Vector of colors for heatmap.}

\item{show_values}{Logical for whether to show correlation values on the heatmap.}

\item{return_matrix}{Return the correlation matrix without plotting correlation heatmap.}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap.}
}
\value{
ggplot2 object of correlation heatmap.
}
\description{
Heatmaps and/or scatter plots to explore replicate concordance of TSSs or TSRs.
}
\details{
Correlation plots are a good way to assess the similarity of samples.
This can be useful to determine replicate concordance and for the initial assessment of
  differences between samples from different conditions.
This function generates various plots using a previously TMM- or MOR-normalized count matrix.
Pearson correlation is recommended for samples from the same technology due to 
the expectation of a roughly linear relationship between the magnitudes of values 
for each feature. Spearman correlation is recommended for samples from different technologies,
such as STRIPE-seq vs. CAGE, due to the expectation of a roughly linear relationship 
between the ranks, rather than the specific values, of each feature.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
exp <- tsr_explorer(TSSs)
exp <- format_counts(exp, data_type="tss")
exp <- normalize_counts(exp, data_type="tss")
plot_correlation(exp, data_type="tss")

}
\seealso{
\code{\link{count_matrix}} to generate the count matrices.
  \code{\link{tmm_normalize}} to TMM normalize the matrices.
}
