% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot Sample Correlation}
\usage{
plot_correlation(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  samples = "all",
  correlation_metric = "pearson",
  log2_transform = TRUE,
  font_size = 12,
  cluster_samples = FALSE,
  heatmap_colors = NULL,
  show_values = TRUE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object with TMM normalized counts}

\item{data_type}{Whether to make scatter plots from TSS, TSR, or RNA-seq & 5' data}

\item{samples}{Either 'all' or the names of the samples to plot}

\item{correlation_metric}{Whether to use Spearman or Pearson correlation}

\item{log2_transform}{Should the TMM values be log2+1 transformed prior to plotting?}

\item{font_size}{The font size for the heatmap tiles}

\item{cluster_samples}{Logical whether hierarchical clustering is performed
on samples in rows and columns.}

\item{heatmap_colors}{Vector of colors for heatmap.}

\item{show_values}{Show the correlation values on the plot.}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap or GGally::ggpairs}
}
\value{
ggplot2 object of correlation plot
}
\description{
Heatmaps and/or scatter plots to explore replicate concordance of TSSs or TSRs.
}
\details{
Correlation plots are a good way to assess the similarity of samples.
This can be useful to compare replicate concordance and for the initial assessment of
  the difference between samples from different conditions or treatments.
This function generates various plots using a previously TMM-normalized count matrix.

'correlation_metric' allows the specification of pearson or spearman correlation
  to be calculated between samples.
Pearson correlation is recommended for samples from the same technology,
  such as RNA-seq, due to the expectation of a roughly linear relationship between 
  the magnitudes of values for each feature.   
Spearmans correlation is recommended for samples from different technologies,
  such as RNA-seq vs. STRIPE-seq, due to the expectation of a roughly linear
  relationship between the ranks, rather than the specific values, of each feature.

'correlation_plot' specifies the type of plot to be generated by the function.
'heatmap', 'scatter', and 'combined' plots are generated by the ggpairs plotting
  function of the GGally library.
The 'hierarchical' plot is generated using the Heatmap function of ComplexHeatmap.

'log2_transform' lets you specify whether you want the data to be log2 transformed
  prior to plotting.
Note that the transformation is done after the correlation metric is calculated,
  and is only meant to facilitate  visualization of the results.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- count_matrix(tsre_exp, data_type="tss")
tsre_exp <- tmm_normalize(tsre_exp, data_type="tss")
plot_correlation(tsre_exp, data_type="tss")

}
\seealso{
\code{\link{count_matrix}} to generate the count matrices.
  \code{\link{tmm_normalize}} to TMM normalize the matrices.
}
