% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plots.R
\name{plot_detected_features}
\alias{plot_detected_features}
\title{Plot Detected Features}
\usage{
plot_detected_features(detected_features, ncol = 1)
}
\arguments{
\item{detected_features}{Tibble of detected feature counts from detect_features}

\item{ncol}{Number of columens to plot data to}
}
\value{
ggplot2 object of detected feature counts
}
\description{
Plot number of features detected per sample.
}
