% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotide.R
\name{plot_dinucleotide_frequencies}
\alias{plot_dinucleotide_frequencies}
\title{Plot Dinucleotide Frequencies}
\usage{
plot_dinucleotide_frequencies(dinucleotide_frequencies)
}
\arguments{
\item{dinucleotide_frequencies}{tibble from dinucleotide_frequencies analysis}
}
\value{
ggplot2 object of dinucleotide frequencies plot
}
\description{
Plot results from dinucleotide analysis
}
