% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tss_plots.R
\name{plot_dominant_tss}
\alias{plot_dominant_tss}
\title{Plot Dominant TSS}
\usage{
plot_dominant_tss(dominant_tss, upstream = 2000, downstream = 500)
}
\arguments{
\item{dominant_tss}{Tibble of dominant TSSs from dominant_tss}

\item{upstream}{Bases upstream to display on plot}

\item{downstream}{Bases downstream to display on plot}
}
\value{
ggplot2 object dominant TSS plot
}
\description{
Plot Dominant TSS
}
