% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(heatmap_matrix, max_value = 5, ncol = 1,
  background_color = "#F0F0F0", low_color = "#56B1F7",
  high_color = "#132B43", ...)
}
\arguments{
\item{heatmap_matrix}{TSS or TSR heatmap matrix from tss_heatmap_matrix ot tsr_heatmap_matrix}

\item{max_value}{Max log2 value to truncate heatmap color}

\item{ncol}{Number of columns when plotting multiple samples}

\item{background_color}{The color of heatmap background}

\item{low_color}{The low value gradient color}

\item{high_color}{The high value gradient color}

\item{...}{Arguments passed to geom_tile}
}
\value{
ggplot2 object of TSS heatmap
}
\description{
Plot heatmap from count matrix generated by tss_heatmap_matrix or tsr_heatmap_matrix
}
