% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  experiment,
  samples = "all",
  data_type = c("tss", "tsr"),
  upstream = 1000,
  downstream = 1000,
  threshold = NULL,
  use_normalized = FALSE,
  dominant = FALSE,
  remove_antisense = TRUE,
  rasterize = FALSE,
  raster_dpi = 150,
  max_value = NULL,
  low_color = "white",
  high_color = "blue",
  log2_transform = TRUE,
  x_axis_breaks = 100,
  ncol = 3,
  filtering = NULL,
  ordering = score,
  order_descending = TRUE,
  order_fun = sum,
  order_samples = NULL,
  quantiling = NULL,
  quantile_fun = sum,
  n_quantiles = 5,
  quantile_samples = NULL,
  split_by = NULL,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{remove_antisense}{Remove antisense reads.}

\item{rasterize}{Rasterize a ggplot.}

\item{raster_dpi}{If rasterization is set, this controls the rasterization DPI.}

\item{max_value}{Truncate heatmap scale at this value.}

\item{low_color}{Color for minimum value.}

\item{high_color}{Color for maximum value.}

\item{log2_transform}{Log2 + 1 transform values for plotting.}

\item{x_axis_breaks}{The distance breaks to show values on the x-axis.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{filtering}{Logical statment to filter data by.}

\item{ordering}{Symbol/name specifying the column to order by.}

\item{order_descending}{Whether to order in descending (TRUE) order.}

\item{order_fun}{Function to aggregate variable by before ordering.}

\item{order_samples}{Samples that are used to calculate ordering.}

\item{quantiling}{Character specifying column to quantile by.}

\item{quantile_fun}{Functiont o aggregate variable by before quantiling.}

\item{n_quantiles}{Number of quantiles.}

\item{quantile_samples}{Samples to use for quantiling.}

\item{split_by}{Named list with split group as name and vector of genes,
or data.frame with columns 'feature' and 'split_group'.}

\item{...}{Additional arguments passed to Heatmap}
}
\value{
ggplot2 object of TSS or TSR heatmap
}
\description{
Plot heatmap from count matrix generated by tss_heatmap_matrix or tsr_heatmap_matrix
}
\details{
This plotting function generates a ggplot2 heatmap of TSS or TSR signal
  surrounding annotated TSSs of genes or transcripts.
Whether genes or transcripts are used depends on the feature type chosen
  when annotating the TSSs with the 'annotate_features' function.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data=annotation,
  data_type="tss", feature_type="transcript"
)
hm_mat <- tss_heatmap_matrix(tsre_exp)
plot_heatmap(hm_matrix)

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
\code{\link{tss_heatmap_matrix}} to generate the TSS matrix data for plotting.
\code{\link{tsr_heatmap_matrix}} to generate the TSR matrix data for plotting.
}
