% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  heatmap_matrix,
  max_value = 5,
  ncol = 1,
  background_color = "#F0F0F0",
  low_color = "#56B1F7",
  high_color = "#132B43",
  ...
)
}
\arguments{
\item{heatmap_matrix}{TSS or TSR heatmap matrix from tss_heatmap_matrix ot tsr_heatmap_matrix}

\item{max_value}{Max log2 (+ 1? qq) value at which to truncate heatmap color}

\item{ncol}{Number of columns to use when plotting multiple samples}

\item{background_color}{The color of the heatmap background (is this the 0 value? qq)}

\item{low_color}{The low value gradient color}

\item{high_color}{The high value gradient color}

\item{...}{Arguments passed to geom_tile}
}
\value{
ggplot2 object of TSS or TSR heatmap
}
\description{
Plot heatmap from count matrix generated by tss_heatmap_matrix or tsr_heatmap_matrix
}
\details{
This plotting function generates a ggplot2 heatmap of TSS or TSR signal
  surrounding annotated TSSs of genes or transcripts.
Whether genes or transcripts are used depends on the feature type chosen
  when annotating the TSSs with the 'annotate_features' function.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data=annotation,
  data_type="tss", feature_type="transcript"
)
hm_mat <- tss_heatmap_matrix(tsre_exp)
plot_heatmap(hm_matrix)

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
\code{\link{tss_heatmap_matrix}} to generate the TSS matrix data for plotting.
\code{\link{tsr_heatmap_matrix}} to generate the TSR matrix data for plotting.
}
