% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  experiment,
  samples = "all",
  data_type = c("tss", "tsr"),
  upstream = 1000,
  downstream = 1000,
  threshold = NULL,
  use_normalized = FALSE,
  dominant = FALSE,
  data_conditions = conditions(data_ordering = ordering(desc(score))),
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{data_conditions}{Apply advanced conditions to the data.}

\item{...}{Additional arguments passed to Heatmap}
}
\value{
ggplot2 object of TSS or TSR heatmap
}
\description{
Plot heatmap from count matrix generated by tss_heatmap_matrix or tsr_heatmap_matrix
}
\details{
This plotting function generates a ggplot2 heatmap of TSS or TSR signal
  surrounding annotated TSSs of genes or transcripts.
Whether genes or transcripts are used depends on the feature type chosen
  when annotating the TSSs with the 'annotate_features' function.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package="TSRexploreR")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data=annotation,
  data_type="tss", feature_type="transcript"
)
hm_mat <- tss_heatmap_matrix(tsre_exp)
plot_heatmap(hm_matrix)

}
\seealso{
\code{\link{annotate_features}} to annotate the TSSs or TSRs.
\code{\link{tss_heatmap_matrix}} to generate the TSS matrix data for plotting.
\code{\link{tsr_heatmap_matrix}} to generate the TSR matrix data for plotting.
}
