% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_metrics.R
\name{plot_max_utr}
\alias{plot_max_utr}
\title{Plot Max UTR Length}
\usage{
plot_max_utr(max_utr, upstream = 1000, downstream = 100, ncol = 1,
  consider_score = FALSE, ...)
}
\arguments{
\item{max_utr}{tibble of max UTRs output by max_utr}

\item{upstream}{Bases upstream to extend average to}

\item{downstream}{Bases downstream to extend average to}

\item{ncol}{Number of columns to plot the data to}

\item{...}{Arguments passed to geom_density}
}
\value{
ggplot2 object of max UTR length average
}
\description{
Plot TSS with furthest distance
}
