% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_num_de}
\alias{plot_num_de}
\title{Plot DE Numbers}
\usage{
plot_num_de(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  de_comparisons = "all",
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The comparisons to plot}

\item{...}{Additional arguments passed to geom_col}
}
\description{
Plot number of DE features.
}
