% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Dimensionality Reduction.}
\usage{
plot_reduction(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  method = "umap",
  n_neighbors = 2,
  min_dist = 0.1,
  use_normalized = TRUE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{method}{'umap' or 'pca'}

\item{n_neighbors}{Nearest neighbors parameter for umap}

\item{min_dist}{Minimum distance between points parameter for umap}

\item{...}{Additional arguments passed to geom_point}
}
\value{
ggplot2 plot of dimension reduction
}
\description{
Dimensionality reduction plot using PCA or UMAP.
}
\details{
This function will generate a dimensionality reduction plot.
These help to visualize the relative similarity of samples
  based on the most variables features.

'method' lets you choose between the more traditional PCA plot or newer
  UMAP plot.

If a UMAP plot is chosen, there are two parameters available
  for tweaking the results.
'n_neighbors' will find the specified number of nearest neighbor points,
  and 'min_dist' specifies the minimum distance allowed between two points 
  in the UMAP plot.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- tss_clustering(tsre_exp)
plot_reduction(tsre_exp, data_type="tsr")

}
