% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Dimensionality Reduction.}
\usage{
plot_reduction(experiment, data_type = c("tss", "tsr", "tss_features",
  "tsr_features"), method = "umap", n_neighbors = 2, min_dist = 0.1,
  ...)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{data_type}{Either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{method}{'umap' or 'pca'}

\item{n_neighbors}{Nearest neighbors parameter for umap}

\item{min_dist}{Minimum distance between points parameter for umap}

\item{...}{Additional arguments passed to geom_point}
}
\value{
ggplot2 plot of dimension reduction
}
\description{
Dimension reduction plot using PCA or UMAP.
}
\details{
This function will generate a dimension reduction plot.
These help to visualize the relative similairty or difference of samples
  based on the most variables features.

'method' lets you choose between the more traditional PCA plot or newer
  UMAP plot.

If a UMAP plot is chosen, there are two parameters available
  for tweaking the results.
'n_neighbors' will find the specified number of nearest neighbor points,
  and 'min_dist' specifies the minimum distance two points are allowed to be
  in space when plotting.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
tsre_exp <- tss_clustering(tsre_exp)
plot_reduction(tsre_exp, data_type = "tsr")

}
