% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{plot_sequence_colormap}
\alias{plot_sequence_colormap}
\title{Plot Sequence Colormap}
\usage{
plot_sequence_colormap(tss_sequences, ncol = 1)
}
\arguments{
\item{tss_sequences}{Sequences surrounding TSS generated with tss_sequences}

\item{ncol}{Number of columns to plot data if quantiles not specified}
}
\value{
ggplot2 object of sequence colormap
}
\description{
Make a sequence colormap for the sequences around TSSs.
}
