% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{plot_sequence_colormap}
\alias{plot_sequence_colormap}
\title{Plot Sequence Colormap}
\usage{
plot_sequence_colormap(tss_sequences, ncol = 1, base_colors = c(A =
  "#109649", C = "#255C99", G = "#F7B32C", T = "#D62839"), text_size = 6,
  ...)
}
\arguments{
\item{tss_sequences}{Sequences surrounding TSS generated with tss_sequences}

\item{ncol}{Number of columns to plot data if quantiles not specified}

\item{base_colors}{Named vector specifying colors for each base}

\item{text_size}{Size of text for plots}

\item{...}{Arguments passed to geom_tile}
}
\value{
ggplot2 object of sequence colormap
}
\description{
Make a sequence colormap for the sequences around TSSs.
}
\details{
This plotting function generates a ggplot2 base color map surrounding TSSs.
Base color maps represent each base surrounding a TSS as a different color.
Since the base composition for every TSS can be seen in one plot, it's a good
  companion figure to sequence logos.

The color of each base is set using the 'base_colors' argument.
The argument input should be a named vector, with the base as the name,
and the desired color of the base as the vector element.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
assembly <- system.file("extdata", "S288C_Assembly.fasta", package = "tsrexplorer")
seqs <- tss_sequences(tsre_exp, genome_assembly = assembly)
plot_sequence_colormap(seqs)

}
\seealso{
\code{\link{tss_sequences}} to get the surrounding sequence.
\code{\link{plot_sequence_logo}} to plot a sequence logo.
}
