% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{plot_sequence_logo}
\alias{plot_sequence_logo}
\title{Generate Sequence Logo}
\usage{
plot_sequence_logo(tss_sequences, ncol = 1, font_size = 6)
}
\arguments{
\item{tss_sequences}{Sequences surrounding TSSs generated with tss_sequences}

\item{ncol}{Number of columns to plot if quantiles is not set}

\item{font_size}{Font size for plots}
}
\value{
ggplot2 object with sequence logo
}
\description{
Create a sequence logo for the sequences around TSSs.
}
\details{
This plotting function uses the ggseqlogo library to make sequence logos
  from the sequences retrieved by the 'tss_sequences' function.
Sequence logos show the enrichment of bases with certain positional biases
  in a centered set of sequences.
This is particularly important for TSS analysis since literature has shown
  strong base preferences spanning TSSs and surrounding sequences.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
assembly <- system.file("extdata", "S288C_Assembly.fasta", package="tsrexplorer")
seqs <- tss_sequences(tsre_exp, genome_assembly=assembly)
plot_sequence_logo(seqs)

}
\seealso{
\code{\link{tss_sequences}} to get the surrounding sequences.
\code{\link{plot_sequence_colormap}} for a sequence color map plot.
}
