% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_count}
\alias{plot_shift_count}
\title{Shift Count Plot}
\usage{
plot_shift_count(
  experiment,
  samples = "all",
  ncol = 3,
  return_table = FALSE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{return_table}{Return a table of results instead of a plot.}

\item{...}{Arguments passed to geom_col}
}
\description{
Shift Count Plot
}
