% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_count}
\alias{plot_shift_count}
\title{Shift Count Plot}
\usage{
plot_shift_count(experiment, samples = "all", ncol = 3)
}
\arguments{
\item{samples}{Samples to plot}

\item{ncol}{Number of columns}

\item{fdr_cutoff}{FDR cutoff value}
}
\description{
Shift Count Plot
}
