% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_count}
\alias{plot_shift_count}
\title{Shift Count Plot}
\usage{
plot_shift_count(
  experiment,
  samples = "all",
  ncol = 3,
  return_table = FALSE,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{return_table}{Return a table of results instead of a plot.}

\item{...}{Arguments passed to geom_col}
}
\value{
ggplot2 of stacked barplot of shifted TSS clusters.
  If 'return_table' is TRUE, a data.frame of underlying counts
  is returned instead.
}
\description{
Generate a stacked barplot for the number of TSS clusters
  shifted upstream and/or downstream per sample.
}
\details{
The 'tss_shifting' function uses the Earth Movers Distance (EMD)
  to find significant upstream or downstream shifts in TSS clusters.
EMD returns a score between -1 and 1 inclusive, with negative and positive scores
  correspending to upstream and downstream shifts respectively.
In rarer cases you can have a shifting scores of 0,
  indicating an equal shift of TSS cluster 'mass' upstream and downstream.

This function generates a stacked barplot for the number of TSS clusters
  with an upstream and downstream shift.
If 'return_table' is TRUE, a data.frame is returned that provides the underlying
  counts for each sample.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "TSRexploreR")
TSSs <- readRDS(TSSs)

tsre <- TSSs[c(1, 4)] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  tss_shift(
    tsre,
    sample_1=c(TSS="S288C_WT_1", TSR="S288C_WT_1"),
    sample_2=c(TSS="S288C_D_1", TSR="S288C_D_1"),
    comparison_name="Untreated_vs_Diamide",
    max_distance = 100, min_threshold = 10, n_resamples = 1000L
  )
\donttest{plot_shift_count(tsre)}

}
\seealso{
\code{\link{tss_shifting}} For TSS cluster shifting calculation.
}
