% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_rank}
\alias{plot_shift_rank}
\title{Shifting Rank Plot}
\usage{
plot_shift_rank(
  experiment,
  samples = "all",
  fdr_cutoff = 0.05,
  score_order = "descending",
  ncol = 3
)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{Shifting results to plot}

\item{fdr_cutoff}{FDR value cutoff}

\item{score_order}{Either descending or ascending}
}
\description{
Shifting Rank Plot
}
