% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_rank}
\alias{plot_shift_rank}
\title{Shifting Rank Plot}
\usage{
plot_shift_rank(
  experiment,
  samples = "all",
  score_order = "descending",
  ncol = 3
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{score_order}{Either 'descending' or 'ascending'.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}
}
\value{
ggplot2 object of ranked shifting scores.
}
\description{
Barplot of shift scores ranked by score.
}
\details{
The 'tss_shifting' function uses the earth mover's score (EMS) to assess shifts
in TSS distribution in consensus TSRs between two samples. This function generates 
a barplot of shifting scores, with the x-axis corresponding to TSS clusters ordered 
by tss shifting score, and the y-axis the shifting score. TSRs can be in ascending 
or descending order as controlled by 'score_order'.
}
\examples{
data(TSSs)
assembly <- system.file("extdata", "S288C_Assembly.fasta", package = "TSRexploreR")
samples <- data.frame(
  sample_name=c(sprintf("S288C_D_\%s", seq_len(3)), sprintf("S288C_WT_\%s", seq_len(3))),
  file_1=rep(NA, 6), file_2=rep(NA, 6),
  condition=c(rep("Diamide", 3), rep("Untreated", 3))
)

exp <- TSSs \%>\%
  tsr_explorer(sample_sheet=samples, genome_assembly=assembly) \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  merge_samples(data_type="tss", merge_group="condition") \%>\%
  merge_samples(data_type="tsr", merge_group="condition") \%>\%
  tss_shift(
    sample_1=c(TSS="S288C_WT_1", TSR="S288C_WT_1"),
    sample_2=c(TSS="S288C_D_1", TSR="S288C_D_1"),
    comparison_name="Untreated_vs_Diamide",
    max_distance = 100, min_threshold = 10, n_resamples = 1000L
  )
  
p <- plot_shift_rank(exp)

}
\seealso{
\code{\link{tss_shift}} to calculate TSS cluster shifting.
}
