% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting_plots.R
\name{plot_shift_rank}
\alias{plot_shift_rank}
\title{Shifting Rank Plot}
\usage{
plot_shift_rank(
  experiment,
  samples = "all",
  score_order = "descending",
  ncol = 3
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{score_order}{Either descending or ascending}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}
}
\value{
ggplot2 object of shifting rank plot.
}
\description{
Barplot of shift ranks ordered by score.
}
\details{
The 'tss_shifting' function uses the Earth Movers Distance (EMD)
  to find significant upstream or downstream shifts in TSS clusters.
EMD returns a score between -1 and 1 inclusive, with negative and positive scores
  correspending to upstream and downstream shifts respectively.
In rarer cases you can have a shifting scores of 0,
  indicating an equal shift of TSS cluster 'mass' upstream and downstream.

This functions generates a barplot of shifting score,
  with the x-axis corresponding to TSS clusters ordered by tss shifting score,
  and the y-axis the shifting score.
TSS clusters can be in ascending or descending order as controlled by 'score_order'.

This function
}
\examples{
data(TSSs)

tsre <- TSSs[c(1, 4)] \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3) \%>\%
  tss_shift(
    sample_1=c(TSS="S288C_WT_1", TSR="S288C_WT_1"),
    sample_2=c(TSS="S288C_D_1", TSR="S288C_D_1"),
    comparison_name="Untreated_vs_Diamide",
    max_distance = 100, min_threshold = 10, n_resamples = 1000L
  )
\donttest{plot_shift_rank(tsre)}

}
\seealso{
\code{\link{tss_shift}} to calculate TSS cluster shifting.
}
