% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{plot_threshold_exploration}
\alias{plot_threshold_exploration}
\title{Plot Threshold Exploration}
\usage{
plot_threshold_exploration(threshold_data, ncol = 1, point_size = 1,
  ...)
}
\arguments{
\item{threshold_data}{Tibble of threshold exploration data from explore_thresholds}

\item{ncol}{Number of columns to plot data}

\item{point_size}{The size of the points on the plot}

\item{...}{Arguments passed to geom_point}
}
\value{
ggplot2 object containing the threshold exploration plot
}
\description{
Make a plot to explore threshold values.
}
\details{
All TSS mapping technologies have some degree of spurious background reads.
tsr_explorer allows for basic thresholding of these background reads by requiring
  a minimum number of reads for a TSS to be considered.

An easy way to pick a threshold with the threshold plot generated
  by 'plot_exploration_threshold'.
This plot shows the proportion of promoter-proximal TSS positions when
  various naive threshold values are tested.
The 'max_threshold' value in the 'explore_thresholds' function defined the maximum threshold
  value tested from 1 to that value.
A recommended final naive threshold value will likely result in a promoter-proximal
  fraction above .80 to .85 and without sacrificing the ability to detect too
  many unique genes or transcripts.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
thresh_results <- explore_thresholds(tsre_exp)
plot_threshold_exploration(thresh_results)

}
\seealso{
\code{\link{explore_thresholds}} for initial calculations.
}
