% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{plot_threshold_exploration}
\alias{plot_threshold_exploration}
\title{Plot Threshold Exploration}
\usage{
plot_threshold_exploration(threshold_data, ncol = 1, point_size = 1,
  sample_order = NULL, ...)
}
\arguments{
\item{threshold_data}{Tibble of threshold exploration data from explore_thresholds}

\item{ncol}{Number of columns to plot data}

\item{point_size}{The size of the points on the plot}

\item{sample_order}{Optional vector specifying sample order in plot}

\item{...}{Arguments passed to geom_point}
}
\value{
ggplot2 object containing the threshold exploration plot
}
\description{
Make a plot to explore threshold values
}
