% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{plot_threshold_exploration}
\alias{plot_threshold_exploration}
\title{Plot Threshold Exploration}
\usage{
plot_threshold_exploration(threshold_data, ncol = 1, point_size = 1,
  ...)
}
\arguments{
\item{threshold_data}{Tibble of threshold exploration data from explore_thresholds}

\item{ncol}{Number of columns in which to plot data}

\item{point_size}{The size of the points on the plot}

\item{...}{Arguments passed to geom_point}
}
\value{
ggplot2 object containing the threshold exploration plot
}
\description{
Make a plot to explore threshold values.
}
\details{
We have found that a threshold of 3 often provides an appropriate balance between
  a high promoter-proximal fraction (>= 0.8) and the number of unique genes or
  transcripts with at least one unique TSS.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
annotation <- system.file("extdata", "S288C_Annotation.gtf", package = "tsrexplorer")
tsre_exp <- annotate_features(
  tsre_exp, annotation_data = annotation,
  data_type = "tss", feature_type = "transcript"
)
thresh_results <- explore_thresholds(tsre_exp)
plot_threshold_exploration(thresh_results)

}
\seealso{
\code{\link{explore_thresholds}} for initial calculations.
}
