% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_plots.R
\name{plot_tsr_metric}
\alias{plot_tsr_metric}
\title{Plot TSR Metrics}
\usage{
plot_tsr_metric(
  experiment,
  tsr_metrics,
  plot_type = "violin",
  samples = "all",
  log2_transform = FALSE,
  ncol = 1,
  use_normalized = FALSE,
  dominant = FALSE,
  threshold = NULL,
  data_conditions = NA,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{tsr_metrics}{Names of metrics in TSRexploreR TSR GRanges to plot}

\item{plot_type}{Output either a 'violin', 'jitter', 'box', or 'boxjitter' plot (default: violin)}

\item{samples}{A vector of sample names to analyze.}

\item{log2_transform}{Whether the metric should be log2 + 1 transformed prior to plotting}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE will discard any TSSs and/or TSRs that do not have the
highest score per gene or transcript.}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}

\item{data_conditions}{Apply advanced conditions to the data.}

\item{...}{Arguments passed to ggplot2 plotting functions}
}
\value{
ggplot2 object with TSR matrix plotted
}
\description{
Plot selected TSR metrics.
}
