% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_plots.R
\name{plot_tsr_metric}
\alias{plot_tsr_metric}
\title{Plot TSR Stats}
\usage{
plot_tsr_metric(experiment, tsr_metrics, samples = "all",
  log2_transform = FALSE, ncol = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with TSR granges}

\item{samples}{Either 'all' or a vector of sample names to analyze}

\item{log2_transform}{Whether the metric should be log2 transformed prior to plotting}

\item{ncol}{Number of columns to plot data}

\item{sample}{Name of sample to return plot for}

\item{tsr_metric}{Names of metrics in tsrexplorer TSR granges to plot}
}
\value{
ggplot2 object with tsr metrix plotted
}
\description{
Plot selected TSR stats
}
