% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_plots.R
\name{plot_tsr_metric}
\alias{plot_tsr_metric}
\title{Plot TSR Metrics}
\usage{
plot_tsr_metric(
  experiment,
  tsr_metrics,
  plot_type = "violin",
  samples = "all",
  log2_transform = FALSE,
  ncol = 1,
  use_normalized = FALSE,
  dominant = FALSE,
  threshold = NULL,
  data_conditions = NULL,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{tsr_metrics}{Names of metrics in TSRexploreR TSR GRanges to plot.}

\item{plot_type}{Output either a 'violin', 'jitter', 'box', or 'boxjitter' plot.}

\item{samples}{A vector of sample names to analyze.}

\item{log2_transform}{Whether the metric should be log2 + 1 transformed prior to plotting.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{data_conditions}{Apply advanced conditions to the data.}

\item{...}{Arguments passed to ggplot2 plotting functions.}
}
\value{
ggplot2 object with TSR matrix plotted.
}
\description{
Plot selected TSR metrics.
}
\details{
Plot any TSR metrics contained within the counts data.table.
Metrics can be supplied as a character vector to 'tsr_metrics',
  and will be optionally Log2 transformed if 'log2_transform' is TRUE.
Valid plot types that can be supplied to 'plot_type' include
  'violin', 'box', 'jitter', and 'boxjitter' (a combination of boxplot and jitterplot).

A set of functions to control data structure for plotting are included. 'use_normalized' 
will use  normalized scores, which only matters if 'consider_score' is TRUE.
'threshold' defines the minimum number of raw counts a TSS or TSR must have to be 
considered. dominant' specifies whether only the dominant TSS or TSR (determined
using the 'mark_dominant' function) is considered. For TSSs, this can be either 
dominant TSS per TSR or gene/transcript, and for TSRs it is the dominant TSR 
per gene/transcript. 'data_conditionals' can be used to filter, quantile, order, 
and/or group data for plotting.
}
\examples{
data(TSSs_reduced)

tsre <- TSSs \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3)
p <- plot_tsr_metric(tsre, "width")

}
\seealso{
\code{\link{tsr_metrics}} to calculate additional TSR metrics.
}
