% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{plot_volcano}
\alias{plot_volcano}
\title{DE Volcano Plot}
\usage{
plot_volcano(differential_expression, log2fc_cutoff = 1,
  fdr_cutoff = 0.05, ...)
}
\arguments{
\item{log2fc_cutoff}{Log2 fold change cutoff for significance}

\item{fdr_cutoff}{FDR value cutoff for significance}

\item{...}{Arguments passed to geom_point}

\item{differential_tsrs}{Tibble of differential TSRs  or genes (RNA-seq)from differential_expression}
}
\value{
ggplot2 object of differential TSRs volcano plot.
}
\description{
Generate volcano plot for differential TSRs or Genes (RNA-seq)
}
