% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{DE Volcano Plot}
\usage{
plot_volcano(
  experiment,
  data_type = c("tss", "tsr", "tss_features", "tsr_features"),
  de_comparisons = "all",
  log2fc_cutoff = 1,
  fdr_cutoff = 0.05,
  ncol = 1,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{either 'tss', 'tsr', 'tss_features', or 'tsr_features'}

\item{de_comparisons}{The DE comparisons to plot}

\item{log2fc_cutoff}{Values below the |Log2 FC| cutoff will be discarded.}

\item{fdr_cutoff}{Values below the FDR cutoff will be discarded.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{...}{Arguments passed to geom_point}
}
\description{
DE Volcano Plot
}
