% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{DE Volcano Plot}
\usage{
plot_volcano(
  experiment,
  data_type = c("tss", "tsr"),
  de_comparisons = "all",
  log2fc_cutoff = 1,
  fdr_cutoff = 0.05,
  ncol = 1,
  ...
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{either 'tss' or 'tsr'.}

\item{de_comparisons}{Character vector of differential expression comparisons to plot.}

\item{log2fc_cutoff}{Differential features not meeting this |Log2 FC| threshold will not be considered.}

\item{fdr_cutoff}{Differential features not meeting this significance threshold will not be considered.}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}

\item{...}{Arguments passed to geom_point.}
}
\value{
ggplot2 object of the volcano plot.
}
\description{
Generate a volcano plot for differential TSSs or TSRs.
}
\details{
This function generates a volcano plot of the results from differential analysis of 
TSSs or TSRs.

'de_comparisons' are the names given to the comparisons from the 'comparison_name'
  argument of the 'differential_expression' function.
'log2fc_cutoff' and 'fdr_cutoff' are the Log2 Fold Change and FDR cutoffs used
  for consideration of significance in the plot.
}
\examples{
data(TSSs)
sample_sheet <- data.frame(
  sample_name=c(
    sprintf("S288C_D_\%s", seq_len(3)),
    sprintf("S288C_WT_\%s", seq_len(3))
  ),
  file_1=rep(NA, 6), file_2=rep(NA, 6),
  condition=c(rep("Diamide", 3), rep("Untreated", 3))
)

tsre <- TSSs \%>\%
  tsr_explorer(sample_sheet=sample_sheet) \%>\%
  format_counts(data_type="tss")

# Differential TSS volcano plot.
diff_tss <- tsre \%>\%
  fit_de_model(data_type="tss", formula= ~condition, method="edgeR") \%>\%
  differential_expression(
    data_type="tss",
    comparison_name="Diamide_vs_Untreated",
    comparison_type="coef",
    comparison=2
  )
p <- plot_volcano(diff_tss, data_type="tss")

}
\seealso{
\code{\link{fit_de_model}} to fit a differential expression model.
\code{\link{differential_expression}} to find differential TSSs or TSRs.
}
