% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditionals.R
\name{quantiling}
\alias{quantiling}
\title{Quantiling}
\usage{
quantiling(
  by = NULL,
  n = NULL,
  samples = NULL,
  descending = TRUE,
  aggr_fun = mean
)
}
\arguments{
\item{by}{Continuous metric for calculating quantiles.}

\item{n}{Number of quantiles to calculate for continuous metric.}

\item{samples}{Samples to use when setting quantiles.}

\item{descending}{Order quantiles in descending order.}

\item{aggr_fun}{Function to aggregate scores if more than one sample selected.}
}
\description{
Quantiling
}
