% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{retrieve_seqs}
\alias{retrieve_seqs}
\title{Retrieve Sequences}
\usage{
retrieve_seqs(
  experiment,
  data_type = c("tss", "tsr", "shift"),
  samples = "all",
  genome_assembly = NULL,
  threshold = NULL,
  dominant = FALSE,
  fixed_size = FALSE,
  extend_upstream = NULL,
  extend_downstream = NULL,
  return_format = "biostrings",
  output_dir = NULL
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{data_type}{Either 'tss', 'tsr', or 'shift'}

\item{samples}{A vector of sample names to analyze.}

\item{genome_assembly}{Genome assembly in FASTA or BSgenome format.}

\item{threshold}{TSSs or TSRs with a score below this value will not be considered.}

\item{dominant}{If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or 
highest-scoring TSR per gene or transcript.}

\item{fixed_size}{Whether all returned sequences should have the
same fixed size starting from the center of the ranges}

\item{extend_upstream}{Distance to extend ranges upstream}

\item{extend_downstream}{Distance to extend ranges downstream}

\item{return_format}{Either 'table' or 'biostrings'}

\item{output_dir}{If set the results will be saved to this directory.
If return format is set to 'biostrings' the sequences will be saved as
FASTA files. If return format is a table it will be saved as a
tab delimited table.}
}
\description{
Retrieve Sequences
}
