% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{ShiftScores}
\alias{ShiftScores}
\title{Shifting Score}
\usage{
ShiftScores(fhash, sample_indicator, distances, scores,
  calc_pvalue = TRUE, nresamp = 100L,
  baseline_level = sample_indicator[1], nthresh = 2,
  check_sort = TRUE)
}
\arguments{
\item{fhash}{fhash of set}

\item{sample_indicator}{The column with the two sample names}

\item{distances}{bin positions}

\item{scores}{bin scores}

\item{calc_pvalue}{Should p-value be returned for comparison}

\item{nresamp}{number of resamples}

\item{baseline_level}{control smaple}

\item{nthresh}{both samples must have at least this number of reads}

\item{check_sort}{Check that the input is sorted properly}
}
\description{
Calculating shifting score and p-value.
}
