% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{ShiftScores}
\alias{ShiftScores}
\title{Shifting Score}
\usage{
ShiftScores(fhash, sample_indicator, distances, scores,
  calc_pvalue = TRUE, nresamp = 100L,
  baseline_level = sample_indicator[1], nthresh = 2,
  check_sort = TRUE)
}
\arguments{
\item{fhash}{fhash of set}

\item{sample_indicator}{The column with the names of the samples to be compared (where is the column from qq)}

\item{distances}{bin positions}

\item{scores}{bin scores}

\item{calc_pvalue}{Whether p-values should be returned for comparisons}

\item{nresamp}{Number of resamplings for the permutation test}

\item{baseline_level}{Control smaple}

\item{nthresh}{Both samples must have at least this number of reads in each TSR}

\item{check_sort}{Check that the input is sorted properly (by fhash? qq)}
}
\description{
Calculate shifting scores and associated permutation test p-values.
}
