% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{ShiftScores}
\alias{ShiftScores}
\title{Shift Score}
\usage{
ShiftScores(
  tss_table,
  baseline_level,
  calc_pvalue = TRUE,
  nresamp = 100L,
  nthresh = 2,
  check_sort = TRUE
)
}
\arguments{
\item{tss_table}{Table of TSSs prepared for shifting score calculation.}

\item{baseline_level}{The sample being used as the baseline for calculation.}

\item{calc_pvalue}{Whether p-values should be returned for comparisons.}

\item{nresamp}{Number of resamplings for permutation test.}

\item{nthresh}{Minimum number of raw counts required in each TSR for both samples.}

\item{check_sort}{Check that the input is sorted properly.}
}
\description{
Calculate shift scores and associated permutation test p-values.
}
