% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softclip_analysis.R
\name{softclip_composition}
\alias{softclip_composition}
\title{Composition of Softclipped Bases}
\usage{
softclip_composition(experiment, samples = "all", n_bases = NULL, ncol = 3)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{n_bases}{Number of bases from -1 position to keep}

\item{ncol}{Integer specifying the number of columns to arrange multiple plots.}
}
\description{
Composition of Softclipped Bases
}
