% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\docType{methods}
\name{tsr_experiment}
\alias{tsr_experiment}
\alias{tsr_experiment,tsr_explorer-method}
\title{Add TSRs}
\usage{
tsr_experiment(tsrexplorer_object)

\S4method{tsr_experiment}{tsr_explorer}(tsrexplorer_object)
}
\arguments{
\item{tsrexplorer_object}{tsr_explorer object}
}
\description{
Convenience function to add TSRs to tsrexplorer object
}
