% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrexplorer.R
\docType{class}
\name{tsr_object-class}
\alias{tsr_object-class}
\title{TSRexplorer Class}
\description{
TSRexplorer Class
}
\section{Slots}{

\describe{
\item{\code{experiment}}{Named lists containing GRanges of TSSs and/or TSRs}

\item{\code{cores}}{Integer value for number of cores available}

\item{\code{annotated}}{Named lists of annotated TSSs and/or TSRs}

\item{\code{counts}}{Named lists of TMM and CPM normalized TSSs and/or TSRs}

\item{\code{correlation}}{Named lists of correlation values between TSS and/or TSR sets}
}}

