% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrexplorer.R
\docType{class}
\name{tsr_explorer-class}
\alias{tsr_explorer-class}
\alias{tsr_explorer}
\title{TSRexplorer Class}
\usage{
tsr_explorer(TSSs = NA, TSRs = NA, cores = 1)
}
\arguments{
\item{TSSs}{Named list of TSS GRanges returned by TSRchitect}

\item{TSRs}{Named list of TSR GRanges returned by TSRchitect}

\item{cores}{Number of CPU cores/threads available}
}
\value{
A tsrexplorer object
}
\description{
TSRexplorer Class

TSRexplorer constructor function.
}
\section{Slots}{

\describe{
\item{\code{experiment}}{Named lists containing GRanges of TSSs and/or TSRs}

\item{\code{cores}}{Integer value for number of cores available}

\item{\code{annotated}}{Named lists of annotated TSSs and/or TSRs}

\item{\code{counts}}{Named lists of TMM and CPM normalized TSSs and/or TSRs}

\item{\code{correlation}}{Named lists of correlation values between TSS and/or TSR sets}
}}

