% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrexplorer.R
\name{tsr_explorer}
\alias{tsr_explorer}
\title{TSRexplorer constructor function.}
\usage{
tsr_explorer(TSSs, TSRs = NA, cores = 1)
}
\arguments{
\item{TSSs}{Named list of TSS GRanges returned by TSRchitect}

\item{TSRs}{Named list of TSR GRanges returned by TSRchitect}

\item{cores}{Number of CPU cores/threads available}
}
\value{
A tsrexplorer object
}
\description{
TSRexplorer constructor function.
}
\examples{
tsrexplorer(TSSs, 4)

}
