% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{tsr_export}
\alias{tsr_export}
\title{Export TSRs}
\usage{
tsr_export(experiment, samples = "all", file_type = "bed",
  out_dir = NA, diff_tsr = FALSE)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{Samples to export ('all' as well? qq)}

\item{file_type}{either 'bed' or 'table'}

\item{out_dir}{Output directory for files}

\item{diff_tsr}{Whether to pull out the differential TSRs (qq again, this is a bit unclear)}
}
\description{
Export TSRs to table or BED
}
