% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{tsr_heatmap_matrix}
\alias{tsr_heatmap_matrix}
\title{TSR Heatmap Count Matrix}
\usage{
tsr_heatmap_matrix(
  experiment,
  samples = "all",
  upstream = 1000,
  downstream = 1000,
  threshold = NA,
  use_normalized = FALSE,
  dominant = FALSE,
  data_conditions = list(order_by = "score")
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}

\item{use_normalized}{Whether to use the normalized (TRUE) or raw (FALSE) counts.}

\item{dominant}{If TRUE will discard any TSSs and/or TSRs that do not have the
highest score per gene or transcript.}

\item{data_conditions}{Apply advanced conditions to the data.}
}
\value{
Matrix of counts for each gene/transcript and position
}
\description{
Generate count matrix to make TSR heatmap
}
