% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tsr_import}
\alias{tsr_import}
\alias{tsr_import,character-method}
\alias{tsr_import,data.frame-method}
\title{Import TSRs}
\usage{
tsr_import(tsrexplorer_obj, object, ...)

\S4method{tsr_import}{character}(tsrexplorer_obj, object)

\S4method{tsr_import}{data.frame}(tsrexplorer_obj, object)
}
\arguments{
\item{tsrexplorer_obj}{tsrexplorer object to which to add TSRs}

\item{object}{Object with TSRs to import into tsrexplorer}

\item{...}{Additional arguments for classes}
}
\value{
tsr explorer object with added TSRs.
}
\description{
Function to import TSRs from various sources.
}
\details{
TSRexploreR can import TSRs from various sources.
Currently beds and TSRchitect tssObjects are supported.

To import beds, a sample sheet must first be created.
The sample sheet can be a data.frame or tabular file.
It should have three columns: sample_name, file_1, and file_2.
sample_name specifies the sample name that will be added for that TSR in the
  tsr explorer object.
file_1 should be the path to the bed file,
  and file_2 can be left blank or filled with NA values.

To import TSRs directly from a TSRchitect tssObject,
  the TSRchitect workflow must first be run up to the
  'determineTSR' step.
}
\seealso{
\code{\link{tss_import}} to import TSSs.
\code{\link{tss_export}} to export TSSs.
\code{\link{tsr_export}} to export TSRs.
}
