% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\docType{methods}
\name{tsr_import}
\alias{tsr_import}
\alias{tsr_import,character-method}
\alias{tsr_import,data.frame-method}
\title{Import TSRs}
\usage{
tsr_import(object, ...)

\S4method{tsr_import}{character}(tsrexplorer_obj, object)

\S4method{tsr_import}{data.frame}(tsrexplorer_obj, object)
}
\arguments{
\item{object}{Object with TSRs to import into tsrexplorer}

\item{...}{Additional arguments for classes}

\item{tsrexplorer_obj}{tsrexplorer object to add the TSRs to}
}
\description{
Convenience function to import TSRs.
}
