% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\docType{methods}
\name{tsr_import,tssObject-method}
\alias{tsr_import,tssObject-method}
\title{TSRchitect object (tssObject)}
\usage{
\S4method{tsr_import}{tssObject}(tsrexplorer_object, object)
}
\description{
TSRchitect object (tssObject)
}
