% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tsr_import}
\alias{tsr_import}
\title{Import TSRs}
\usage{
tsr_import(experiment, sample_sheet = NULL, file_type = "auto", delim = "\\t")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{file_type}{Either 'auto', 'table', or 'bed'.}

\item{delim}{Delimiter for tabular input.}
}
\value{
TSRexploreR object with added TSRs.
}
\description{
Function to import TSRs from various sources.
}
\details{
TSRexploreR can import TSRs from various sources. Currently, these are BED files,
delimited tables, and TSRchitect tssObjects.

To import TSRs from files, a sample sheet must first be created. The sample sheet can 
be a data.frame or tabular file. It should have three columns: sample_name, file_1, 
and file_2. sample_name specifies the sample name that will be added for that 
TSR in the TSRexploreR object. file_1 should be the path to the file, and 
file_2 can be left blank or filled with NA values.

To import TSRs directly from a TSRchitect tssObject, the TSRchitect workflow must 
first be run up to the 'determineTSR' step.
}
\examples{
\dontrun{
data(TSSs_reduced)

# Export bed as example data.
tsre <- TSSs \%>\%
  tsr_explorer \%>\%
  format_counts(data_type="tss") \%>\%
  tss_clustering(threshold=3)
tsr_export(tsre)

# Import the previously created bed.
samples <- data.frame(
  sample_name="S288C_D_1",
  file_1="S288C_D_1.bed",
  file_2=NA
)
tsr <- tsr_explorer(sample_sheet=samples)
tsr <- tsr_import(tsr)
}

}
\seealso{
\code{\link{tss_import}} to import TSSs.
\code{\link{tss_export}} to export TSSs.
\code{\link{tsr_export}} to export TSRs.
}
