% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_clustering.R
\name{tss_clustering}
\alias{tss_clustering}
\title{TSS Clustering}
\usage{
tss_clustering(experiment, samples = "all", threshold = 1,
  max_distance = 25)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{The samples to call TSRs for}

\item{threshold}{Minimum number of reads for a TSS to be considered}

\item{max_distance}{The maximum distance to cluster TSSs}
}
\value{
tsr explorer object with TSRs
}
\description{
Basic distance and threshold based clustering of TSSs.
}
\details{
This function clusters TSSs into Transcription Start Regions (TSRs).

TSSs are clustered if their score is greater than or equal to
  'threshold' and are less than or equal to 'max_distance'
  from eachother.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
tsre-exp <- tss_clustering(tsre_exp)

}
