% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_clustering.R
\name{tss_clustering}
\alias{tss_clustering}
\title{TSS Clustering}
\usage{
tss_clustering(experiment, samples = "all", threshold = 1, max_distance = 25)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{Samples for which TSSs should be clustered}

\item{threshold}{Consider only TSSs with at least this number of raw counts}

\item{max_distance}{Maximum distance between TSSs that can be clustered}
}
\value{
tsrexplorer object with TSRs
}
\description{
Basic distance and threshold-based clustering of TSSs.
}
\details{
This function clusters TSSs into Transcription Start Regions (TSRs).

TSSs are clustered if their score is greater than or equal to
  'threshold' and are less than or equal to 'max_distance'
  from each other.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
tsre_exp <- tss_clustering(tsre_exp)

}
