% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_clustering.R
\name{tss_clustering}
\alias{tss_clustering}
\title{TSS Clustering}
\usage{
tss_clustering(
  experiment,
  samples = "all",
  threshold = NULL,
  max_distance = 25,
  max_width = NULL
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{threshold}{TSSs or TSRs with a score below this value will be discarded.}

\item{max_distance}{Maximum distance between TSSs that can be clustered}

\item{max_width}{Maximum width of TSR allowed.}
}
\value{
TSRexploreR object with TSRs
}
\description{
Basic distance and threshold-based clustering of TSSs.
}
\details{
This function clusters TSSs into Transcription Start Regions (TSRs).

TSSs are clustered if their score is greater than or equal to
  'threshold' and are less than or equal to 'max_distance'
  from each other.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tsre_exp <- tss_clustering(tsre_exp)

}
