% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_methods.R
\name{tss_experiment}
\alias{tss_experiment}
\alias{tss_experiment,tsr_explorer-method}
\title{Add TSSs}
\usage{
tss_experiment(TSRexploreR_object)

\S4method{tss_experiment}{tsr_explorer}(TSRexploreR_object)
}
\arguments{
\item{TSRexploreR_object}{TSRexploreR object}
}
\description{
Convenience function to add TSSs to a TSRexploreR object.
}
