% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{tss_export}
\alias{tss_export}
\title{Export TSSs}
\usage{
tss_export(
  experiment,
  samples = "all",
  file_type = "bedgraph",
  out_dir = NA,
  diff_tss = FALSE,
  sep = "\\t",
  genome_assembly = NULL
)
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{samples}{A vector of sample names to analyze.}

\item{file_type}{either 'bedgraph', 'table', or 'bigwig'}

\item{out_dir}{Output directory for files}

\item{diff_tss}{If TRUE will output differential TSSs}

\item{sep}{Delimiter for table output}

\item{genome_assembly}{Genome assembly in either a FASTA file or BSgenome object format.}
}
\value{
Either bedgraphs split by strand, or a tabular file.
}
\description{
Export TSSs to tables or bedgraphs
}
\details{
This function will save TSSs as bedgraphs, or a tab delimited file.

'file_type' controls what the TSSs will be output as.
'bedgraph' will result in each sample being saved as two bedgraph files,
  one for each strand.
'table' will output a file with the delimiter specified by the 'sep' argument.
The resulting table will have all columns added to the TSSs data
  in the tsr explorer object, such as various metrics.

The directory to output the files can be set with 'out_dir'.
  A value of NA will save the files to the working directory by default.

If 'diff_tss' is TRUE, only differential TSSs will be output.
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package="TSRexploreR")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type="tss")
tss_export(tsre_exp)

}
\seealso{
\code{\link{tsr_export}} to export TSRs.
\code{\link{tss_import}} to import TSSs.
\code{\link{tsr_import}} to import TSRs.
}
