% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{tss_export}
\alias{tss_export}
\title{Export TSSs}
\usage{
tss_export(experiment, samples = "all", file_type = "bedgraph",
  out_dir = NA)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{samples}{Samples to export}

\item{file_type}{either 'bedgraph' or 'table'}

\item{out_dir}{Output directory for files}
}
\description{
Export TSSs to table or bedgraph
}
