% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{tss_heatmap_matrix}
\alias{tss_heatmap_matrix}
\title{TSS Heatmap Count Matrix}
\usage{
tss_heatmap_matrix(experiment, samples = "all", upstream = 1000,
  downstream = 1000, threshold = NA, use_cpm = FALSE,
  dominant = FALSE, data_conditions = list(order_by = "score"))
}
\arguments{
\item{experiment}{tsrexplorer object with annotated TSSs}

\item{samples}{Either 'all' or name of samples to analyze}

\item{upstream}{Bases upstream to consider}

\item{downstream}{bases downstream to consider}

\item{threshold}{Reads required per TSS}

\item{use_cpm}{Whether to use the CPM normalized values or not}

\item{dominant}{Consider only dominant}

\item{data_conditions}{Condition the data (filter, order, and quantile/group available)}
}
\value{
matrix of counts for each gene/transcript and position
}
\description{
Generate count matrix to make TSS heatmap
}
