% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tss_import}
\alias{tss_import}
\alias{tss_import,character-method}
\alias{tss_import,data.frame-method}
\alias{tss_import,tssObject-method}
\alias{tss_import,CAGEexp-method}
\title{Import TSSs}
\usage{
tss_import(tsrexplorer_obj, object, ...)

\S4method{tss_import}{character}(tsrexplorer_obj, object, sep = "\\t", col_names = TRUE)

\S4method{tss_import}{data.frame}(tsrexplorer_obj, object)

\S4method{tss_import}{tssObject}(tsrexplorer_obj, object)

\S4method{tss_import}{CAGEexp}(tsrexplorer_obj, object, data_type)
}
\arguments{
\item{object}{Object with TSSs to import into tsrexplorer}

\item{...}{Additional arguments for classes}

\item{sep}{Delimiter for sample_sheet file}

\item{col_names}{Logical, whether the sample_sheet file has a header (TRUE) or not (FALSE)}

\item{data_type}{Either "tss", "tsr", or "consensus"}

\item{tsrexplorer}{tsrexplorer object}
}
\value{
tsr explorer object with imported TSSs
}
\description{
Function to import TSSs from various sources.
}
\details{
TSRexploreR can import TSSs from various sources.
Currently bedgraphs, bigwigs, and TSRchitect tssObjects are supported.

To import bedgraphs or bigwigs, a sample sheet must first be created.
The sample sheet can be a data.frame or tabular file.
It should have three columns: sample_name, file_1, and file_2.
sample_name specifies the sample name that will be added for that TSS in the
  tsr explorer object.
file_1 and file_2 should be the path to the two bedgraphs or bigwigs,
  which are usually separated by strand.

To import TSSs directly from a TSRchitect tssObject,
  the TSRchitect workflow must first be run up to the
  'processTSS' step.
}
\seealso{
\code{\link{tsr_import}} to import TSRs.
\code{\link{tss_export}} to export TSSs.
\code{\link{tsr_export}} to export TSRs.
}
