% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\docType{methods}
\name{tss_import}
\alias{tss_import}
\alias{tss_import,character-method}
\alias{tss_import,data.frame-method}
\alias{tss_import,tssObject-method}
\alias{tss_import,CAGEexp-method}
\title{Import TSSs}
\usage{
tss_import(object, ...)

\S4method{tss_import}{character}(object)

\S4method{tss_import}{data.frame}(object, sample_name = "sample")

\S4method{tss_import}{tssObject}(object)

\S4method{tss_import}{CAGEexp}(object, data_type)
}
\arguments{
\item{object}{Object with TSSs to import into tsrexplorer}

\item{...}{Additional arguments for classes}

\item{sample_name}{The name of the sample}
}
\description{
Convenience function to import TSSs.
}
