% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\docType{methods}
\name{tss_import}
\alias{tss_import}
\alias{tss_import,character-method}
\alias{tss_import,data.frame-method}
\alias{tss_import,tssObject-method}
\alias{tss_import,CAGEexp-method}
\title{Import TSSs}
\usage{
tss_import(tsrexplorer_obj, object, ...)

\S4method{tss_import}{character}(tsrexplorer_obj, object, delim = "\\t",
  col_names = TRUE)

\S4method{tss_import}{data.frame}(tsrexplorer_obj, object)

\S4method{tss_import}{tssObject}(tsrexplorer_obj, object)

\S4method{tss_import}{CAGEexp}(tsrexplorer_obj, object, data_type)
}
\arguments{
\item{object}{Object with TSSs to import into tsrexplorer}

\item{...}{Additional arguments for classes}

\item{delim}{Delimiter for sample_sheet file}

\item{col_names}{Logical, whether the sample_sheet file has a header (TRUE) or not (FALSE)}

\item{data_type}{Either "tss", "tsr", or "consensus"}

\item{tsrexplorer}{tsrexplorer object}
}
\description{
Convenience function to import TSSs.
}
