% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tss_import}
\alias{tss_import}
\title{Import TSSs}
\usage{
tss_import(experiment, sample_sheet = NULL, file_type = "auto", delim = "\\t")
}
\arguments{
\item{experiment}{TSRexploreR object.}

\item{sample_sheet}{A sample sheet data.frame or tab delimited file.
Must have the columns 'sample_name', 'file_1', and 'file_2'.
Additional meta-data columns can be added with sample information such as condition and batch.}

\item{file_type}{Either 'auto', 'bedgraph', 'bigwig', 'table', or 'ctss'.}

\item{delim}{Delimiter for tabular input.}
}
\value{
TSRexploreR object with imported TSSs.
}
\description{
Function to import TSSs from various sources.
}
\details{
TSRexploreR can import TSSs from various file formats. Currently, these are bedGraphs, 
bigWigs, CTSS files, delimited tables, and TSRchitect tssObjects.

To import TSSs from files, a sample sheet must first be created. The sample 
sheet can be a data.frame or tabular file. It should have three columns: 
sample_name, file_1, and file_2. sample_name specifies the sample name that 
will be added for that TSS in the TSRexploreR object. For bedGraphs and bigWigs,
file_1 and file_2 should be the paths to the plus and minus strand TSS files. For
CTSS or tabular import, file_1 should be the path to the file, and file_2can be 
left blank or filled with NA values.

To import TSSs directly from a TSRchitect tssObject, the TSRchitect workflow 
must first be run up to the 'processTSS' step.
}
\seealso{
\code{\link{tsr_import}} to import TSRs.
\code{\link{tss_export}} to export TSSs.
\code{\link{tsr_export}} to export TSRs.
}
