% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{tss_normalization}
\alias{tss_normalization}
\title{TMM Normalize TSSs}
\usage{
tss_normalization(experiment)
}
\arguments{
\item{experiment}{tsrexplorer object}
}
\value{
tibble of TMM normalized read counts
}
\description{
Using edgeR to TMM normalize TSSs
}
