% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{tss_sequences}
\alias{tss_sequences}
\title{Retrieve Sequences Near TSSs}
\usage{
tss_sequences(experiment, sample, genome_assembly, threshold = 1,
  distance = 10)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS GRanges}

\item{sample}{Name of sample to analyze}

\item{genome_assembly}{Genome assembly in fasta format}

\item{distance}{Bases to add on each side of TSS}
}
\value{
Sequences surrounding TSSs
}
\description{
Retrieve sequences surrounding TSSs for further plotting
}
