% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{tss_sequences}
\alias{tss_sequences}
\title{Retrieve Sequences Near TSSs}
\usage{
tss_sequences(experiment, samples = "all", genome_assembly,
  threshold = 1, distance = 10, quantiles = 1)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS GRanges}

\item{genome_assembly}{Genome assembly in fasta format}

\item{distance}{Bases to add on each side of TSS}

\item{quantiles}{Break data into quantiles}

\item{sample}{Either "all" or names of samples to analyze}
}
\value{
Sequences surrounding TSSs
}
\description{
Retrieve sequences surrounding TSSs for further plotting
}
