% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{tss_sequences}
\alias{tss_sequences}
\title{Retrieve Sequences Near TSSs}
\usage{
tss_sequences(experiment, samples = "all", genome_assembly,
  threshold = 1, distance = 10, dominant = FALSE,
  data_conditions = list(order_by = "score"))
}
\arguments{
\item{experiment}{tsrexplorer object with TSS GRanges}

\item{samples}{Either "all" or names of samples to analyze}

\item{genome_assembly}{Genome assembly in fasta format or bioconductor BSgenome}

\item{threshold}{Keep only TSSs with threshold number of reads or more}

\item{distance}{Bases to add on each side of TSS}

\item{dominant}{Whether only dominant should be considered}

\item{data_conditions}{Condition the data (filter, quantile, and grouping available)}
}
\value{
DataFrame of sequences surrounding TSSs.
}
\description{
Retrieve sequences surrounding TSSs for further plotting
}
\examples{
TSSs <- system.file("extdata", "S288C_TSSs.RDS", package = "tsrexplorer")
TSSs <- readRDS(TSSs)
tsre_exp <- tsr_explorer(TSSs)
tsre_exp <- format_counts(tsre_exp, data_type = "tss")
assembly <- system.file("extdata", "S288C_Assembly.fasta", package = "tsrexplorer")
seqs <- tss_sequences(tsre_exp, genome_assembly = assembly)

}
\seealso{
\code{\link{plot_sequence_logo}} to make sequence logos.
\code{\link{plot_sequence_colormap}} to make sequence color maps.
}
