% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{tss_sequences}
\alias{tss_sequences}
\title{Retrieve Sequences Near TSSs}
\usage{
tss_sequences(experiment, samples = "all", genome_assembly,
  threshold = 1, distance = 10, dominant = FALSE,
  data_conditions = NA)
}
\arguments{
\item{experiment}{tsrexplorer object with TSS GRanges}

\item{samples}{Either "all" or names of samples to analyze}

\item{genome_assembly}{Genome assembly in fasta format or bioconductor BSgenome}

\item{threshold}{Keep only TSSs with threshold number of reads or more}

\item{distance}{Bases to add on each side of TSS}

\item{dominant}{Whether only dominant should be considered}

\item{data_condtions}{Condition the data (filter, quantile, and grouping available)}
}
\value{
Sequences surrounding TSSs
}
\description{
Retrieve sequences surrounding TSSs for further plotting
}
