% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{tss_shift}
\alias{tss_shift}
\title{TSS Shifting}
\usage{
tss_shift(experiment, compare_samples, min_distance = 100)
}
\arguments{
\item{experiment}{tsrexplorer object}

\item{compare_samples}{Vector of two samples to compare}
}
\description{
Calculate TSS shifting statistics
}
