% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_shifting.R
\name{tss_shift}
\alias{tss_shift}
\title{TSS Shifting}
\usage{
tss_shift(
  experiment,
  sample_1,
  sample_2,
  min_distance = 100,
  min_threshold = 10,
  n_resamples = 1000L
)
}
\arguments{
\item{experiment}{TSRexploreR object}

\item{sample_1}{First sample to compare.
Vector with sample name for TSS and TSR,
with names 'TSS' and 'TSR'}

\item{sample_2}{Second sample to compare.
Vector with sample name for TSS and TSR,
with names 'TSS' and 'TSR'}

\item{min_distance}{TSRs less than this distance apart will be merged}

\item{min_threshold}{Minimum number of raw counts required in each TSR for both TSR samples}

\item{n_resamples}{Number of resamplings for permutation test}
}
\description{
Calculate TSS shifting statistics
}
